% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{add_growth_death_rate}
\alias{add_growth_death_rate}
\title{Add growth/Death rates}
\usage{
add_growth_death_rate(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
df dataframe
}
\description{
Add growth/Death rates
}
