% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{add_iso}
\alias{add_iso}
\title{Add XXX.csv file}
\usage{
add_iso(x, iso, ds, level, map = c("id"), append = TRUE)
}
\arguments{
\item{x}{\code{data.frame} generated by a data source function.}

\item{iso}{ISO code (3 letters).}

\item{ds}{name of the data source function generating \code{x}.}

\item{level}{integer. Granularity level. 1: country-level data. 2: state-level data. 3: city-level data.}

\item{map}{named vector mapping the columns of \code{x} to the columns of the XXX.csv file.}

\item{append}{logical. Append the data to the XXX.csv file if it already exists? Default \code{TRUE}.}
}
\value{
\code{data.frame}
}
\description{
Add new country in the inst/extdata/db/ folder
}
\keyword{internal}
