% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_src}
\alias{add_src}
\title{Add Data Source}
\usage{
add_src(...)
}
\arguments{
\item{...}{named arguments corresponding to the columns of the \href{https://github.com/covid19datahub/COVID19/tree/master/inst/extdata/src.csv}{src.csv} file.}
}
\value{
\code{data.frame}.
}
\description{
Add data source to the file inst/extdata/src.csv
}
\examples{
\dontrun{

 x <- add_src(
  iso_alpha_3 = "USA", 
  administrative_area_level = 1, 
  data_type = "confirmed", 
  url = "https://example.com", 
  title = "New Data Source", 
  year = 2020)
  
}

}
