% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_admin.ch.R
\name{admin.ch}
\alias{admin.ch}
\title{Federal Office of Public Health}
\source{
https://www.covid19.admin.ch/en/overview
}
\usage{
admin.ch(level, state = NULL)
}
\arguments{
\item{level}{1, or 2 (only for Switzerland)}

\item{state}{one of CH (Switzerland) or FL (Liechtenstein)}
}
\description{
Data source for: Switzerland and Liechtenstein
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\keyword{internal}
