% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_AFG.R
\name{AFG}
\alias{AFG}
\title{Afghanistan}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_AFG.R}
}
\usage{
AFG(level)
}
\description{
Afghanistan
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
confirmed cases,
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/AFG.csv}{Ministry of Public Health Afghanistan (2020)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_humdata.af.R}{Afghanistan Ministry of Health}:
confirmed cases,
deaths,
recovered,
tests.
}
}
}

\concept{Level 1}
\concept{Level 2}
