% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggr_to_cont}
\alias{aggr_to_cont}
\title{Aggregates data to continent or subcontinent (group)}
\usage{
aggr_to_cont(data, group, time, popdata, allstatuses)
}
\arguments{
\item{data}{data.frame aggregated data per Country.Region}

\item{group}{character continent or subcontinent}

\item{time}{character date or contagion_day}

\item{popdata}{data with population info}

\item{allstatuses}{character vector of statuses to base the recomputation from: confirmed recovered deaths active}

\item{cont}{character continent name used to subset popdata}
}
\value{
data.frame aggregated at group level
}
\description{
Aggregates data to continent or subcontinent (group)
}
\note{
growth and mortality variables must be recomputed after the aggregation
the return dataset renames the group column into Country.region to allow the following graphs to work
}
