% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{aggregate_country_data}
\alias{aggregate_country_data}
\title{Aggregate data by country}
\usage{
aggregate_country_data(data)
}
\arguments{
\item{data}{data.frameing}
}
\value{
country_df tibble of by country confirmed, deaths, active and recovered for each day
}
\description{
Aggregate data by country
}
