% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{aggregate_province_timeseries_data}
\alias{aggregate_province_timeseries_data}
\title{Aggregate data by province}
\usage{
aggregate_province_timeseries_data(data)
}
\arguments{
\item{data}{data.frameing}
}
\value{
df tibble of by country confirmed, deaths, active and recovered for each day
}
\description{
Aggregate data by province
}
