% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{align_country_names_pop}
\alias{align_country_names_pop}
\title{align countries names between population dataframe and orig_data dataframe}
\usage{
align_country_names_pop(data)
}
\arguments{
\item{data}{data.frame}
}
\description{
align countries names between population dataframe and orig_data dataframe
}
