% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_arcgis.de.R
\name{arcgis.de}
\alias{arcgis.de}
\title{Robert Koch Institute}
\source{
https://www.arcgis.com/home/item.html?id=f10774f1c63e40168479a1feb6c7ca74
}
\usage{
arcgis.de(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Germany
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\keyword{internal}
