% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_arcgis.se.R
\name{arcgis.se}
\alias{arcgis.se}
\title{Public Health Agency of Sweden}
\source{
https://www.arcgis.com/home/item.html?id=b5e7488e117749c19881cce45db13f7e
}
\usage{
arcgis.se(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Sweden
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
}
}

\keyword{internal}
