% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_ARG.R
\name{ARG}
\alias{ARG}
\title{Argentina}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_ARG.R}
}
\usage{
ARG(level)
}
\description{
Argentina
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
confirmed cases,
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ARG.csv}{Statoids (2010)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gob.ar.R}{Argentine Ministry of Health}:
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ARG.csv}{Statoids (2010)}, \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ARG.csv}{Wikipedia (2010)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gob.ar.R}{Argentine Ministry of Health}:
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
