% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_AUS.R
\name{AUS}
\alias{AUS}
\title{Australia}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_AUS.R}
}
\usage{
AUS(level)
}
\description{
Australia
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.m3it.covid19data.R}{Matt Bolton}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care,
patients requiring ventilation.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.m3it.covid19data.R}{Matt Bolton}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care,
patients requiring ventilation.
}
}
}

\concept{Level 1}
\concept{Level 2}
