% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_AUT.R
\name{AUT}
\alias{AUT}
\title{Austria}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_AUT.R}
}
\usage{
AUT(level)
}
\description{
Austria
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gv.at.R}{Federal Ministry of Social Affairs, Health, Care and Consumer Protection, Austria}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gv.at.R}{Federal Ministry of Social Affairs, Health, Care and Consumer Protection, Austria}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gv.at.R}{Federal Ministry of Social Affairs, Health, Care and Consumer Protection, Austria}:
confirmed cases,
deaths,
recovered.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
