% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{backgroud_map_col}
\alias{backgroud_map_col}
\title{background color of maps}
\format{
An object of class \code{character} of length 1.
}
\usage{
backgroud_map_col
}
\description{
background color of maps
}
\keyword{datasets}
