% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{barplots_colors}
\alias{barplots_colors}
\title{Color Palette for simple barplots}
\format{
An object of class \code{list} of length 7.
}
\usage{
barplots_colors
}
\description{
Color Palette for simple barplots
}
\keyword{datasets}
