% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{blank_plot}
\alias{blank_plot}
\title{draw a blank plot with message}
\usage{
blank_plot(where = "", what = "", add = "")
}
\arguments{
\item{where}{character where data are missing}

\item{what}{character variable name, "" if unknown}

\item{add}{character additional text message}
}
\value{
ggplot plot
}
\description{
draw a blank plot with message
}
