% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_BRA.R
\name{BRA}
\alias{BRA}
\title{Brazil}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_BRA.R}
}
\usage{
BRA(level, ...)
}
\description{
Brazil
}
\section{Data Sources}{

\subsection{Level 1}{

\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.wcota.covid19br.R}{Wesley Cota}:
confirmed cases, deaths, recovered, tests, total vaccine doses administered.
}


\subsection{Level 2}{

\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.wcota.covid19br.R}{Wesley Cota}:
confirmed cases, deaths, recovered, tests, total vaccine doses administered.
}


\subsection{Level 3}{

\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.wcota.covid19br.R}{Wesley Cota}:
confirmed cases, deaths.

\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.eguidotti.covid19br.R}{Emanuele Guidotti}:
total vaccine doses administered, people with at least one vaccine dose, people fully vaccinated.
}
}

\concept{level 1}
\concept{level 2}
\concept{level 3}
