% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{breaks_lab}
\alias{breaks_lab}
\title{Utility for labels breaks in plots}
\usage{
breaks_lab(x, breaks = .breaks.xaxis)
}
\arguments{
\item{x}{vector of data}
}
\value{
x breaks values
}
\description{
Utility for labels breaks in plots
}
