% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_data_aggr}
\alias{build_data_aggr}
\title{Builds dataset to be used in modules merging pop_data with data}
\usage{
build_data_aggr(data, popdata)
}
\arguments{
\item{data}{data}

\item{popdata}{population data with continent info}
}
\value{
data
}
\description{
Builds dataset to be used in modules merging pop_data with data
}
