% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_CAN.R
\name{CAN}
\alias{CAN}
\title{Canada}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/iso_CAN.R}
}
\usage{
CAN(level)
}
\description{
Canada
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_canada.ca.R}{Public Health Agency of Canada}:
confirmed cases,
deaths,
recovered,
tests.
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_ourworldindata.org.R}{Our World in Data}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_canada.ca.R}{Public Health Agency of Canada}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}
}

\concept{Level 1}
\concept{Level 2}
