% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_canada.ca.R
\name{canada.ca}
\alias{canada.ca}
\title{Public Health Agency of Canada}
\source{
https://www.canada.ca/en/public-health/services/diseases/coronavirus-disease-covid-19.html
}
\usage{
canada.ca(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Canada
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
