% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{caption_death_fun}
\alias{caption_death_fun}
\title{caption death}
\usage{
caption_death_fun(var)
}
\arguments{
\item{var}{character growth factor variable}
}
\value{
character text for caption
}
\description{
caption death
}
