% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{caption_fitted}
\alias{caption_fitted}
\title{caption fitted values}
\usage{
caption_fitted(col = "Grey", type = "dashed")
}
\arguments{
\item{col}{character color name}

\item{type}{character linetype}
}
\value{
character text for caption
}
\description{
caption fitted values
}
