% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{caption_growth_factor_fun}
\alias{caption_growth_factor_fun}
\title{caption growth factor}
\usage{
caption_growth_factor_fun(growthvar)
}
\arguments{
\item{growthvar}{character growth factor variable}
}
\value{
character text for caption
}
\description{
caption growth factor
}
