% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{caption_source_vaccines}
\alias{caption_source_vaccines}
\title{caption vaccination source}
\usage{
caption_source_vaccines()
}
\value{
character text for caption
}
\description{
caption vaccination source
}
