% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{caption_tests}
\alias{caption_tests}
\title{caption tests}
\usage{
caption_tests(text = "Tests")
}
\arguments{
\item{text}{character text of variable}
}
\value{
character text for caption
}
\description{
caption tests
}
