% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{case_colors_labs}
\alias{case_colors_labs}
\title{Color Palette for Variable labels}
\usage{
case_colors_labs(cc = .case_colors)
}
\arguments{
\item{cc}{vector \code{.case_colors}}
}
\description{
Color Palette for Variable labels
}
