% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_cdc.gov.R
\name{cdc.gov}
\alias{cdc.gov}
\title{Centers for Disease Control and Prevention}
\source{
https://data.cdc.gov/browse?q=COVID-19\%20Vaccination&sortBy=relevance
}
\usage{
cdc.gov(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: United States
}
\section{Level 1}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 2}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 3}{

\itemize{
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
