% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH}
\alias{CH}
\title{Datasets}
\format{An object of class \code{data.frame} with 27 rows and 11 columns.}
\source{
https://www.bfs.admin.ch/bfs/en/home/statistics/regional-statistics/regional-portraits-key-figures/cantons/data-explanations.html
}
\usage{
CH
}
\description{
Datasets
}
\details{
population age is 0-19 not 0-14
}
\keyword{datasets}
