% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_CHE.R
\name{CHE}
\alias{CHE}
\title{Switzerland}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_CHE.R}
}
\usage{
CHE(level)
}
\description{
Switzerland
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{Swiss Federal Statistical Office (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_admin.ch.R}{Federal Office of Public Health}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
recovered.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/CHE.csv}{Swiss Federal Statistical Office (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_admin.ch.R}{Federal Office of Public Health}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.openzh.covid19.R}{Specialist Unit for Open Government Data Canton of Zurich}:
recovered,
patients requiring ventilation.
}
}
}

\concept{Level 1}
\concept{Level 2}
