% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check}
\alias{check}
\title{Internal function - Check}
\usage{
check(x)
}
\arguments{
\item{x}{\code{data.frame}}
}
\value{
\code{data.frame} of \code{logical}.
Rows: country/state/city.
Columns: Variables.
\code{TRUE}: ok.
\code{FALSE}: missing values.
}
\description{
Identify which variables are missing for a given country/state/city.
}
\examples{
\dontrun{
x <- world(raw = TRUE)
y <- COVID19:::check(x)
View(y)
}

}
\keyword{internal}
