% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_flag}
\alias{check_flag}
\title{Check whether data are available}
\usage{
check_flag(data, var)
}
\arguments{
\item{data}{data.frame}

\item{var}{character string, variable name}
}
\value{
logical
}
\description{
Check whether data are available
}
