% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compare_nth_cases_plot.R
\name{choice_nthcases_plot}
\alias{choice_nthcases_plot}
\title{given arguments select choices for nth plot}
\usage{
choice_nthcases_plot(
  vars = vars_nthcases_plot,
  actives = TRUE,
  tests = FALSE,
  hosp = FALSE,
  strindx = FALSE,
  vax = FALSE,
  log = TRUE
)
}
\arguments{
\item{vars}{variable names in the drop down option.}

\item{actives}{if TRUE then add new_active and active variables to vars.}

\item{tests}{if TRUE then add new_test and test variables to vars.}

\item{hosp}{if TRUE then add new_hosp and hosp variables to vars.}

\item{strindx}{if TRUE then add stringency_index variable to vars.}

\item{log}{if FALSE tthen negative variables and new variables are removed}
}
\description{
given arguments select choices for nth plot
}
