% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_plotly_leg}
\alias{clean_plotly_leg}
\title{clean ggplotly legend}
\usage{
clean_plotly_leg(.plotly_x, .extract_str)
}
\arguments{
\item{.plotly_x}{ggplotly object}

\item{.extract_str}{regular expression}
}
\description{
clean ggplotly legend
}
