% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_COL.R
\name{COL}
\alias{COL}
\title{Colombia}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_COL.R}
}
\usage{
COL(level)
}
\description{
Colombia
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gov.co.R}{Instituto Nacional de Salud}:
confirmed cases,
deaths,
recovered,
tests.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gov.co.R}{Instituto Nacional de Salud}:
confirmed cases,
deaths,
recovered,
tests.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gov.co.R}{Instituto Nacional de Salud}:
confirmed cases,
deaths,
recovered,
people with at least one vaccine dose,
people fully vaccinated.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
