% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{combine_hospvars_lev2}
\alias{combine_hospvars_lev2}
\title{replace hospital data of level1 with level2 data}
\usage{
combine_hospvars_lev2(data1, data2, country = "Swizterland")
}
\arguments{
\item{data1}{data.frame level 1 data}

\item{data2}{data.frame level 2 data}

\item{country}{character country name for replacement}
}
\value{
data1 with hospital data from data2
}
\description{
replace hospital data of level1 with level2 data
}
\details{
at level1 hospital data are not complete
}
