% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_group_plots_server.R
\name{confirmed_info}
\alias{confirmed_info}
\title{info text on confirmed cases}
\usage{
confirmed_info(sep = "<br/>")
}
\arguments{
\item{sep}{separator, break line.}
}
\description{
info text on confirmed cases
}
