% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus.R
\name{corona}
\alias{corona}
\title{corona}
\usage{
corona()
}
\description{
Live dataset for coronavirus
}
\examples{
library(ggplot2)
aus = subset(corona(), region == "Australia")
ggplot(aus, aes(x = date, y = active)) +
geom_point() + facet_wrap( ~ state) +
geom_line() + theme_bw()
}
