% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\docType{package}
\name{covid19-package}
\alias{covid19}
\alias{covid19-package}
\title{covid19: A COVID19 data analyser}
\description{
The R package covid19 is used to retrieve, analyse and represent
    Italian data on the COVID19 outbreak and spreading.
}
\author{
\strong{Maintainer}: Michele Stefano Altieri \email{msa@hypermynds.com}

}
\keyword{internal}
