% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{covid19_counties}
\alias{covid19_counties}
\title{Counties}
\usage{
covid19_counties(parse = TRUE)
}
\arguments{
\item{parse}{Logical, defaults to TRUE, indicating whether to convert
response object into a tibble data frame}
}
\value{
A tibble
}
\description{
Reported number of cases of COVID-19 by U.S. county
}
