% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{covid19_parse}
\alias{covid19_parse}
\title{Parser}
\usage{
covid19_parse(x)
}
\arguments{
\item{x}{Input data returned by covid19 function}
}
\value{
A tibble data frame
}
\description{
Parses data returned by covid19 endpoint functions
}
