% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{covid19_screenshots}
\alias{covid19_screenshots}
\title{Screenshots}
\usage{
covid19_screenshots(parse = TRUE)
}
\arguments{
\item{parse}{Logical, defaults to TRUE, indicating whether to convert
response object into a tibble data frame}
}
\value{
A tibble
}
\description{
Screenshot records for reported number of cases of COVID-19 in the U.S.
}
