% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{covid19_states}
\alias{covid19_states}
\title{States}
\usage{
covid19_states(parse = TRUE)
}
\arguments{
\item{parse}{Logical, defaults to TRUE, indicating whether to convert
response object into a tibble data frame}
}
\value{
A tibble
}
\description{
Reported number cases of COVID-19 by state
}
