% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{covid19_states_daily}
\alias{covid19_states_daily}
\title{States daily}
\usage{
covid19_states_daily(parse = TRUE)
}
\arguments{
\item{parse}{Logical, defaults to TRUE, indicating whether to convert
response object into a tibble data frame}
}
\value{
A tibble
}
\description{
Daily reported number of COVID-19 cases by U.S. state
}
