% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{covid19_states_info}
\alias{covid19_states_info}
\title{States info}
\usage{
covid19_states_info(parse = TRUE)
}
\arguments{
\item{parse}{Logical, defaults to TRUE, indicating whether to convert
response object into a tibble data frame}
}
\value{
A tibble
}
\description{
State information about COVID-19
}
