% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{covid19_us}
\alias{covid19_us}
\title{U.S.}
\usage{
covid19_us(parse = TRUE)
}
\arguments{
\item{parse}{Logical, defaults to TRUE, indicating whether to convert
response object into a tibble data frame}
}
\value{
A tibble
}
\description{
Reported number of cases of COVID-19 in the U.S.
}
