% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{covid19_us_daily}
\alias{covid19_us_daily}
\title{U.S. Daily}
\usage{
covid19_us_daily(parse = TRUE)
}
\arguments{
\item{parse}{Logical, defaults to TRUE, indicating whether to convert
response object into a tibble data frame}
}
\value{
A tibble
}
\description{
Daily reported number of cases of COVID-19 in the United States
}
