% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19cite.R
\name{covid19cite}
\alias{covid19cite}
\title{Datasets Citation}
\usage{
covid19cite(x, bibtex = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{dataset generated by the \link{covid19} function.}

\item{bibtex}{logical. Bibtex output? Default \code{FALSE}.}

\item{verbose}{logical. Print citations? Default \code{TRUE}.}
}
\value{
\code{data.frame} of data sources used by the \link{covid19} function.
}
\description{
Cite the data sources used by the \link{covid19} function.
}
\examples{
\dontrun{

# Worldwide data by country
x   <- covid19()
cit <- covid19cite(x)
View(cit)

}

}
