% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_covid19india.org.R
\name{covid19india.org}
\alias{covid19india.org}
\title{COVID-19 India API}
\source{
https://data.covid19india.org
}
\usage{
covid19india.org(level)
}
\arguments{
\item{level}{2}
}
\description{
Data source for: India
}
\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
