% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidData.R
\name{covidData}
\alias{covidData}
\title{covidData}
\usage{
covidData()
}
\value{
dataframe of name of countries, last update,
number of deaths and infections
}
\description{
Extract all data from John Hopkins Repository until today
}
\examples{
df <- covidData()
}
