% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_covidtracking.com.R
\name{covidtracking.com}
\alias{covidtracking.com}
\title{The COVID Tracking Project}
\source{
https://covidtracking.com/data/api
}
\usage{
covidtracking.com(level)
}
\arguments{
\item{level}{2}
}
\description{
Data source for: United States
}
\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\keyword{internal}
