% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_covidtracking_com.R
\name{covidtracking_com}
\alias{covidtracking_com}
\title{The COVID Tracking Project}
\source{
https://covidtracking.com/data/api
}
\usage{
covidtracking_com(level)
}
\description{
Imports confirmed cases, deaths, recovered, tests, and hospitalization data for U.S. states from
The COVID Tracking Project. The Project has ended all data collection as of March 7, 2021.
The existing API will continue to work until May 2021, but will only include data up to March 7, 2021.
}
\keyword{internal}
