% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{data_to_series}
\alias{data_to_series}
\title{Custom and faster version of highcharter:::data_to_series}
\usage{
data_to_series(data, type)
}
\description{
It should be only applied for grouped data containing date column.
}
