% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{datasource}
\alias{datasource}
\title{THIS IS A TEMPLATE TO SUPPORT ADDITIONAL DATA SOURCES}
\usage{
datasource(cache, ...)
}
\description{
Copy this template in a new file
Rename the file with the name of the data source
Rename this function with the name of the data source
}
\details{
The function must include the argument 'cache'. See:
https://github.com/emanuele-guidotti/COVID19/blob/master/R/covid19.R
The function can include additional arguments
The function must return a data.frame

Examples:
https://github.com/emanuele-guidotti/COVID19/blob/master/R/openZH.R
https://github.com/emanuele-guidotti/COVID19/blob/master/R/pcmdpc.R
https://github.com/emanuele-guidotti/COVID19/blob/master/R/jhuCSSE.R
}
\keyword{'date'.}
\keyword{Create}
\keyword{Download}
\keyword{Extra}
\keyword{Include}
\keyword{additional}
\keyword{and}
\keyword{but}
\keyword{cache}
\keyword{clean}
\keyword{code}
\keyword{column}
\keyword{columns,}
\keyword{data,}
\keyword{data.}
\keyword{etc.}
\keyword{filter}
\keyword{here...}
\keyword{https://github.com/emanuele-guidotti/COVID19#csv-data-files}
\keyword{including}
\keyword{internal}
\keyword{limited}
\keyword{not}
\keyword{return}
\keyword{the}
\keyword{to:}
\keyword{your}
