% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\docType{data}
\name{confirmed_timeseries_csv_url}
\alias{confirmed_timeseries_csv_url}
\alias{deaths_timeseries_csv_url}
\alias{recovered_timeseries_csv_url}
\alias{daily_url}
\title{Data Urls confirmed_timeseries}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
confirmed_timeseries_csv_url

deaths_timeseries_csv_url

recovered_timeseries_csv_url

daily_url
}
\description{
Data Urls confirmed_timeseries

Data Urls deaths_timeseries_csv_url

Data Urls recovered_timeseries_csv_url

Data Urls daily_url
}
\keyword{datasets}
