% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{date_bar_plot}
\alias{date_bar_plot}
\title{Evolution by_date}
\usage{
date_bar_plot(df)
}
\arguments{
\item{df}{data.frame with column called Date and x column to plot}
}
\value{
barplot of given variable by contagion
}
\description{
Evolution by_date
}
\examples{
\dontrun{
df <- get_timeseries_full_data() \%>\%
 get_timeseries_by_contagion_day_data() \%>\%
 aggregate_province_timeseries_data() \%>\%
 filter(Country.Region == "Italy")
 select(Country.Region, date, confirmed)

 df <- df \%>\%
 mutate(status = as.factor(Country.Region)) \%>\%
 mutate(value = confirmed) \%>\%
 capitalize_names_df()

df \%>\% date_bar_plot()
}

}
