% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{decreasing}
\alias{decreasing}
\title{Decreasing counts}
\usage{
decreasing(x, k = 1, strict = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{k}{vector of lags (e.g., \code{k=1:7} returns \code{TRUE} if an element is decreasing with respect to any of the previous 7 elements)}

\item{strict}{logical indicating whether only strictly decreasing counts should be considered}
}
\value{
\code{logical} indicating decreasing elements
}
\description{
Check which elements in a numeric vector are decreasing with respect to the previous elements
}
\keyword{internal}
