% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{delay_date}
\alias{delay_date}
\title{determine latest date by subtracting from today the below computation}
\usage{
delay_date()
}
\description{
determine latest date by subtracting from today the below computation
}
