% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy}
\alias{deploy}
\title{Helper function to deploy up-to-date data}
\usage{
deploy()
}
\description{
Download and deploy most recent data. 
    Will only work if \link{rsconnect} information has been set up.
}
