% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond.R
\name{diamond}
\alias{diamond}
\title{Diamond Princess}
\usage{
diamond(raw = FALSE)
}
\arguments{
\item{raw}{logical. Skip data cleaning? Default \code{FALSE}}
}
\value{
Grouped \code{tibble} (\code{data.frame}). \href{https://github.com/emanuele-guidotti/COVID19}{More info}
}
\description{
Tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.
Diamond Princess data.
}
\details{
See \href{https://github.com/emanuele-guidotti/COVID19}{data sources}
}
\examples{
x <- diamond()

}
