% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_DNK.R
\name{DNK}
\alias{DNK}
\title{Denmark}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/iso_DNK.R}
}
\usage{
DNK(level)
}
\description{
Denmark
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
confirmed cases,
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_ssi.dk.R}{Statens Serum Institut}:
confirmed cases,
deaths,
tests,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_ssi.dk.R}{Statens Serum Institut}:
confirmed cases,
tests,
people with at least one vaccine dose,
people fully vaccinated.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
