% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{docstring}
\alias{docstring}
\title{Generate docstring for a data source}
\usage{
docstring(ds, name, ...)
}
\arguments{
\item{ds}{the name of the ds_ R function}

\item{name}{the name of the data provider}

\item{...}{arguments passed to the ds_ function}
}
\description{
Generate docstring for a data source
}
\keyword{internal}
