% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{.breaks.xaxis}
\alias{.breaks.xaxis}
\title{length breaks of y axis}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.breaks.xaxis
}
\description{
length breaks of y axis
}
\keyword{datasets}
