% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.case_colors}
\alias{.case_colors}
\title{Color Palette for Areaplot variables}
\format{
An object of class \code{character} of length 5.
}
\usage{
.case_colors
}
\description{
Color Palette for Areaplot variables
}
\keyword{datasets}
