% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.current_vars}
\alias{.current_vars}
\title{Variables defining current state}
\format{
An object of class \code{character} of length 10.
}
\usage{
.current_vars
}
\description{
Variables defining current state
}
\keyword{datasets}
