% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getContinents}
\alias{.getContinents}
\title{Global definition of continent names}
\usage{
.getContinents(idx, contname)
}
\arguments{
\item{idx}{character, name of continent to be extracted: "Europe", "Asia", "Africa", "LatAm & Carib.", "Northern America", "Oceania"}
}
\value{
data.frame with continent info
}
\description{
Global definition of continent names
}
