% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.hosp_colors}
\alias{.hosp_colors}
\title{Color Palette Hospitalised vars}
\format{
An object of class \code{character} of length 4.
}
\usage{
.hosp_colors
}
\description{
Color Palette Hospitalised vars
}
\keyword{datasets}
