% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.hosp_vars}
\alias{.hosp_vars}
\title{Variables related to hospitalization}
\format{
An object of class \code{character} of length 2.
}
\usage{
.hosp_vars
}
\description{
Variables related to hospitalization
}
\keyword{datasets}
