% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.hosp_vars_datahub}
\alias{.hosp_vars_datahub}
\title{Variables related to hospitalisation in covid19datahub}
\format{
An object of class \code{character} of length 3.
}
\usage{
.hosp_vars_datahub
}
\description{
Variables related to hospitalisation in covid19datahub
}
\keyword{datasets}
