% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.index_vars}
\alias{.index_vars}
\title{Variables defined as 0-100 index in map plot}
\format{
An object of class \code{character} of length 1.
}
\usage{
.index_vars
}
\description{
Variables defined as 0-100 index in map plot
}
\keyword{datasets}
