% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.neg_vars}
\alias{.neg_vars}
\title{Variables where negative values are allowed in map plot}
\format{
An object of class \code{character} of length 10.
}
\usage{
.neg_vars
}
\description{
Variables where negative values are allowed in map plot
}
\keyword{datasets}
