% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.rate_vars}
\alias{.rate_vars}
\title{Variables defined as rate in map plot}
\format{
An object of class \code{character} of length 4.
}
\usage{
.rate_vars
}
\description{
Variables defined as rate in map plot
}
\keyword{datasets}
