% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_country.R
\docType{data}
\name{.Selected_Country}
\alias{.Selected_Country}
\title{Default country, to open first in the page}
\format{
An object of class \code{character} of length 1.
}
\usage{
.Selected_Country
}
\description{
Default country, to open first in the page
}
\keyword{datasets}
