% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.tests_colors}
\alias{.tests_colors}
\title{Color Palette Tests vars}
\format{
An object of class \code{character} of length 2.
}
\usage{
.tests_colors
}
\description{
Color Palette Tests vars
}
\keyword{datasets}
