% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.vacc_colors}
\alias{.vacc_colors}
\title{Color Palette vaccination vars}
\format{
An object of class \code{character} of length 5.
}
\usage{
.vacc_colors
}
\description{
Color Palette vaccination vars
}
\keyword{datasets}
