% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.vacc_vars}
\alias{.vacc_vars}
\title{Variables related to hospitalization in datahub}
\format{
An object of class \code{character} of length 5.
}
\usage{
.vacc_vars
}
\description{
Variables related to hospitalization in datahub
}
\keyword{datasets}
