% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_all_data}
\alias{download_all_data}
\title{Download all data for COVID-19 dashboard}
\usage{
download_all_data(path = NULL)
}
\arguments{
\item{path}{file download path}
}
\description{
Download all data for COVID-19 dashboard
}
