% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_google_mr_us}
\alias{download_google_mr_us}
\title{Download and tidy the United States Google Mobility Report data}
\usage{
download_google_mr_us()
}
\value{
A data table containing the data.
}
\description{
Download and tidy the United States Google Mobility Report data
}
