% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_jhu_covid_19}
\alias{download_jhu_covid_19}
\title{Download and tidy the Johns Hopkins University COVID-19 data for the United States}
\usage{
download_jhu_covid_19()
}
\value{
A data table containing the data.
}
\description{
Download and tidy the Johns Hopkins University COVID-19 data for the United States
}
