% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_usda}
\alias{download_usda}
\title{Download and tidy the US Department of Agriculture data for Education, Unemployment/Income, and Population}
\usage{
download_usda()
}
\arguments{
\item{path}{File download path}
}
\description{
Download and tidy the US Department of Agriculture data for Education, Unemployment/Income, and Population
}
