% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_DPC.R
\name{DPC}
\alias{DPC}
\title{Cruise Ship Diamond Princess}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_DPC.R}
}
\usage{
DPC(level)
}
\description{
Cruise Ship Diamond Princess
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
confirmed cases,
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_wikipedia.dp.R}{Wikipedia}:
confirmed cases,
tests.
}
}
}

\concept{Level 1}
