% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{drop_decreasing}
\alias{drop_decreasing}
\title{Drop decreasing counts}
\usage{
drop_decreasing(x, by, cols, k, strict)
}
\arguments{
\item{x}{\code{data.frame} containing the column \code{date}}

\item{by}{vector to group by}

\item{cols}{vector of columns to clean}

\item{k}{vector of lags (e.g., \code{k=1:7} removes values that are decreasing with respect to any of the previous 7 values)}

\item{strict}{logical indicating whether only strictly decreasing counts should be dropped}
}
\value{
\code{x} where decreasing counts are replaced with \code{NA}
}
\description{
Drop decreasing counts
}
\keyword{internal}
