% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_docstring}
\alias{ds_docstring}
\title{Generate docstring to use in the ds_ files}
\usage{
ds_docstring(ds, name, desc, url, ...)
}
\arguments{
\item{ds}{the name of the ds_ R function}

\item{name}{the name of the data provider}

\item{desc}{the name(s) of the countries supported by the data provider, e.g., "United States".}

\item{url}{the link to the data provider}

\item{...}{arguments passed to the ds_ function}
}
\description{
Generate docstring to use in the ds_ files
}
\keyword{internal}
