% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{err_log}
\alias{err_log}
\title{Error Log}
\usage{
err_log(x)
}
\arguments{
\item{x}{\code{data.frame} generated by the function \code{\link{covid19}}}
}
\value{
\code{data.frame}
}
\description{
Error log of the \code{\link{covid19}} output.
}
\examples{
\dontrun{

x <- covid19(level = 2)
e <- err_log(x)

}

}
