% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{extdata}
\alias{extdata}
\title{External Data}
\usage{
extdata(...)
}
\arguments{
\item{...}{path to file}
}
\value{
\code{data.frame}
}
\description{
Read files in the inst/extdata/ folder.
}
\keyword{internal}
