% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{fetch_data}
\usage{
fetch_data()
}
\value{
A dataframe
}
\description{
this function fetches corona data from John Hopkins University database and updates it every day.
}
\examples{
fetch_data()
}
