% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fix_colors}
\alias{fix_colors}
\title{Fix colors}
\usage{
fix_colors(p, labs = FALSE)
}
\arguments{
\item{p}{ggplot object}

\item{labs}{logical, if TRUE then variables labels arte used}
}
\value{
p ggplot object
}
\description{
Fix colors
}
