% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fix_legend_position}
\alias{fix_legend_position}
\title{Fix legend position}
\usage{
fix_legend_position(p)
}
\arguments{
\item{p}{ggplot object}
}
\value{
p ggplot object
}
\description{
Fix legend position
}
