% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\docType{data}
\name{French.Colonies}
\alias{French.Colonies}
\title{named vector of french colonies to be added to France}
\format{
An object of class \code{character} of length 6.
}
\usage{
French.Colonies
}
\description{
named vector of french colonies to be added to France
}
\keyword{datasets}
