% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_plots.R
\name{from_contagion_day_bar_facet_plot}
\alias{from_contagion_day_bar_facet_plot}
\title{Evolution from contagion day facet}
\usage{
from_contagion_day_bar_facet_plot(df)
}
\arguments{
\item{df}{data.frame to plot}
}
\value{
barplot facet
}
\description{
Evolution from contagion day facet
}
