% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{funformat}
\alias{funformat}
\title{Format in ggplotly}
\usage{
funformat(x, perc, digits = NULL)
}
\arguments{
\item{x}{numeric vector of data}

\item{perc}{logical data are percentage if TRUE}
}
\value{
x formatted x
}
\description{
Format in ggplotly
}
