% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_text}
\alias{gen_text}
\title{Derives text for plotly popups}
\usage{
gen_text(x, namvar)
}
\arguments{
\item{x}{data}

\item{namvar}{character vector variable name}
}
\value{
text
}
\description{
Derives text for plotly popups
}
