% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_geohive.ie.R
\name{geohive.ie}
\alias{geohive.ie}
\title{Health Protection Surveillance Centre (HPSC) and Health Service Executive (HSE)}
\source{
https://covid-19.geohive.ie/search
}
\usage{
geohive.ie(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Ireland
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\keyword{internal}
