% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aggrvars}
\alias{get_aggrvars}
\title{Global definition of numeric aggregatable vars in dataset}
\usage{
get_aggrvars()
}
\value{
character vector
}
\description{
Global definition of numeric aggregatable vars in dataset
}
