% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Covid19_version.R
\name{get_Covid19_version}
\alias{get_Covid19_version}
\title{Covid19 version}
\usage{
get_Covid19_version()
}
\value{
The version of Covid19 as character string.
}
\description{
Returns the version of the installed Covid19 package and app.
}
