% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cumvars}
\alias{get_cumvars}
\title{Global definition of numeric cumulative vars in dataset}
\usage{
get_cumvars()
}
\value{
character vector
}
\description{
Global definition of numeric cumulative vars in dataset
}
