% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_daily_data}
\alias{get_daily_data}
\title{Get daily data}
\usage{
get_daily_data(date)
}
\arguments{
\item{date}{string of format mm-dd-yyyy}
}
\value{
data data.frame
}
\description{
Get daily data
}
