% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_datahub}
\alias{get_datahub}
\title{Get timeseries full data from datahub}
\usage{
get_datahub(country = NULL, stardate = "2020-01-22", lev = 1, verbose = FALSE)
}
\arguments{
\item{country}{character country, to chose with lev = 2}

\item{stardate}{character staring date}

\item{lev}{integer 1 for country level, 2 for reagions}

\item{verbose}{logical. Print data sources? Default FALSE (opposite from \code{covid19})}
}
\value{
data tibble of confirmed, deaths, active and recovered Country.Region
}
\description{
Get timeseries full data from datahub
}
\details{
data sourced from https://github.com/covid19datahub/COVID19/
}
