% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_date_data}
\alias{get_date_data}
\title{date data}
\usage{
get_date_data(data, date)
}
\arguments{
\item{data}{data.frame}

\item{date}{date format yyyy-mm-dd}
}
\value{
global tibble of confirmed, deaths, active and recovered for each day
}
\description{
date data
}
