% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_nytimes.R
\name{get_nytimes_state_col_types}
\alias{get_nytimes_state_col_types}
\title{Returns a \code{col_spec} for NY Times COVID-19 data.}
\usage{
get_nytimes_state_col_types()
}
\value{
\code{col_spec} to be passed to \code{readr::read_csv} for the NY Times COVID-19 case dataset.
}
\description{
Returns a \code{col_spec} for NY Times COVID-19 data to override the defaults detected
by \code{readr::read_csv}.
}
