% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_owid_jhcsse}
\alias{get_owid_jhcsse}
\title{Import JHU data from Our World in Data}
\usage{
get_owid_jhcsse(
 
    url = "https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/jhu/full_data.csv"
)
}
\description{
Import JHU data from Our World in Data
}
