% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_pop_data}
\alias{get_pop_data}
\title{population data}
\usage{
get_pop_data(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
global tibble of confirmed, deaths, active and recovered for each day by population
}
\description{
population data
}
\examples{
\dontrun{
orig_data <- get_timeseries_full_data() \%>\%
              get_timeseries_by_contagion_day_data() \%>\%
              select(-ends_with("rate"))
data <- orig_data \%>\% align_country_names()

}

}
