% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_pop_datahub}
\alias{get_pop_datahub}
\title{retrieves population data, matches countries with continents and subcontinents in line with get_datahub}
\usage{
get_pop_datahub()
}
\value{
global tibble of confirmed, deaths, active and recovered for each day by population
}
\description{
retrieves population data, matches countries with continents and subcontinents in line with get_datahub
}
\note{
additional demographic variables are available for further development
}
\examples{
\dontrun{
orig_data <- get_timeseries_full_data() \%>\%
              get_timeseries_by_contagion_day_data()
data <- orig_data \%>\% align_country_names()

}

}
