% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state_populations.R
\name{get_state_populations_col_types}
\alias{get_state_populations_col_types}
\title{Returns a \code{col_spec} for U.S. Census Bureau estimated state population data.}
\usage{
get_state_populations_col_types()
}
\value{
\code{col_spec} to be passed to \code{readr::read_csv} for the U.S. Census Bureau
estimated state population dataset.
}
\description{
Returns a \code{col_spec} for U.S. Census Bureau estimated state population data
to override the defaults detected by \code{readr::read_csv}.
}
