% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries_by_contagion_day_data}
\alias{get_timeseries_by_contagion_day_data}
\title{Get data by day contagion}
\usage{
get_timeseries_by_contagion_day_data(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data tibble of confirmed, deaths, active and recovered, each for Province.State, Country.Region, Lat, Long, contagion_day and day
}
\description{
Get data by day contagion
}
