% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries_country_data}
\alias{get_timeseries_country_data}
\title{Country.Region data timeseries}
\usage{
get_timeseries_country_data(data, country)
}
\arguments{
\item{data}{data.frame}

\item{country}{name of the country. Character string}
}
\value{
country_df tibble of by country confirmed, deaths, active and recovered for each day
}
\description{
Country.Region data timeseries
}
