% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries_data}
\alias{get_timeseries_data}
\title{Get timeseries data}
\usage{
get_timeseries_data()
}
\value{
data list of confirmed, deaths and recovered, each for Province.State, Country.Region, Lat, Long, and day
}
\description{
Get timeseries data
}
