% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries_full_data}
\alias{get_timeseries_full_data}
\title{Get timeseries full data}
\usage{
get_timeseries_full_data()
}
\value{
data tibble of confirmed, deaths, active and recovered, each for Province.State, Country.Region, Lat, Long, and day
}
\description{
Get timeseries full data
}
