% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries_global_data}
\alias{get_timeseries_global_data}
\title{Global data timeseries}
\usage{
get_timeseries_global_data(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
global tibble of global confirmed, deaths, active and recovered, for each day
}
\description{
Global data timeseries
}
