% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries_province_data}
\alias{get_timeseries_province_data}
\title{Province.State data timeseries}
\usage{
get_timeseries_province_data(data, province)
}
\arguments{
\item{data}{data.frame}

\item{province}{name of the province Character string}
}
\value{
province_df tibble of by country confirmed, deaths, active and recovered for each day
}
\description{
Province.State data timeseries
}
