% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_timeseries_single_data}
\alias{get_timeseries_single_data}
\title{Get timeseries single data}
\usage{
get_timeseries_single_data(param)
}
\arguments{
\item{type}{data to be returned. Character string. Expected values c("confirmed", "deaths", "recovered")}
}
\value{
data data.frame for Province.State, Country.Region, Lat, Long, and day
}
\description{
Get timeseries single data
}
