% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{get_trends_data_new}
\alias{get_trends_data_new}
\title{Compute case and deaths trends}
\usage{
get_trends_data_new(df, time_unit_extent = 14)
}
\arguments{
\item{df}{JHSU CSSE data}

\item{time_unit_extent}{defaults to 14 days}
}
\value{

}
\description{
Compute case and deaths trends
}
