% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getdg_lab}
\alias{getdg_lab}
\title{Derives number of digits for rounding}
\usage{
getdg_lab(dg, maxv)
}
\arguments{
\item{dg}{integer number of characters of figure, say 1000 = 4}

\item{maxv}{numeric max value}
}
\value{
integer number of digit
}
\description{
Derives number of digits for rounding
}
