% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTableOptions}
\alias{getTableOptions}
\title{Get table options.}
\usage{
getTableOptions(scrollX = TRUE, maxrowsperpage = 5)
}
\arguments{
\item{scrollX}{Param to allow scrollX. Default TRUE}

\item{maxrowsperpage}{Maximum number of rows to display per page. Default 5.}
}
\description{
Returns option list for datatable.
}
