% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.ceedsdemm.covidprodataset.R
\name{github.ceedsdemm.covidprodataset}
\alias{github.ceedsdemm.covidprodataset}
\title{Centre of Excellence in Economics and Data Science, University of Milan}
\source{
https://github.com/CEEDS-DEMM/COVID-Pro-Dataset
}
\usage{
github.ceedsdemm.covidprodataset(level)
}
\arguments{
\item{level}{3}
}
\description{
Data source for: Italy
}
\section{Level 3}{

\itemize{
\item deaths
}
}

\keyword{internal}
