% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.eguidotti.covid19br.R
\name{github.covid19datahub.covid19br}
\alias{github.covid19datahub.covid19br}
\title{Emanuele Guidotti}
\source{
https://github.com/covid19datahub/covid19br
}
\usage{
github.covid19datahub.covid19br(level)
}
\arguments{
\item{level}{3}
}
\description{
Data source for: Brazil
}
\section{Level 3}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
