% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.covid19euzh.covid19eudata.R
\name{github.covid19euzh.covid19eudata}
\alias{github.covid19euzh.covid19eudata}
\title{COVID-19 European Chinese Channel}
\source{
https://github.com/covid19-eu-zh/covid19-eu-data
}
\usage{
github.covid19euzh.covid19eudata(level)
}
\arguments{
\item{level}{2}
}
\description{
Data source for: Poland
}
\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item tests
}
}

\keyword{internal}
