% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.cssegisanddata.covid19.R
\name{github.cssegisanddata.covid19}
\alias{github.cssegisanddata.covid19}
\title{Johns Hopkins Center for Systems Science and Engineering}
\source{
https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data/csse_covid_19_time_series
}
\usage{
github.cssegisanddata.covid19(
  level = 1,
  file = "global",
  country = NULL,
  state = NULL
)
}
\arguments{
\item{level}{1, 2, or 3 (U.S only)}

\item{file}{one of "global" for worldwide data or "US" for U.S. data}

\item{country}{filter by name of country}

\item{state}{filter by name of state}
}
\description{
Data source for: Worldwide
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\section{Level 3}{

\itemize{
\item confirmed cases (U.S. only)
\item deaths (U.S. only)
}
}

\keyword{internal}
