% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.cssegisanddata.covid19unified.R
\name{github.cssegisanddata.covid19unified}
\alias{github.cssegisanddata.covid19unified}
\title{Johns Hopkins Center for Systems Science and Engineering}
\source{
https://github.com/CSSEGISandData/COVID-19_Unified-Dataset
}
\usage{
github.cssegisanddata.covid19unified(level, iso)
}
\arguments{
\item{level}{1, 2, 3}

\item{iso}{the 2-letter (level=1) or 3-letter (level>1) ISO code of the country}
}
\description{
Data source for: Worldwide
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\keyword{internal}
