% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.dsfsi.covid19za.R
\name{github.dsfsi.covid19za}
\alias{github.dsfsi.covid19za}
\title{Data Science for Social Impact research group, University of Pretoria}
\source{
https://github.com/dsfsi/covid19za
}
\usage{
github.dsfsi.covid19za(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: South Africa
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
}
}

\keyword{internal}
