% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.dssgpt.covid19ptdata.R
\name{github.dssgpt.covid19ptdata}
\alias{github.dssgpt.covid19ptdata}
\title{Data Science for Social Good Portugal}
\source{
https://github.com/dssg-pt/covid19pt-data
}
\usage{
github.dssgpt.covid19ptdata(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Portugal
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
}
}

\keyword{internal}
