% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.italia.covid19opendatavaccini.R
\name{github.italia.covid19opendatavaccini}
\alias{github.italia.covid19opendatavaccini}
\title{Commissario straordinario per l'emergenza Covid-19, Presidenza del Consiglio dei Ministri}
\source{
https://github.com/italia/covid19-opendata-vaccini
}
\usage{
github.italia.covid19opendatavaccini(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Italy
}
\section{Level 1}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 2}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
