% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.jmcastagnetto.covid19perudata.R
\name{github.jmcastagnetto.covid19perudata}
\alias{github.jmcastagnetto.covid19perudata}
\title{Jesus M. Castagnetto}
\source{
https://github.com/jmcastagnetto/covid-19-peru-data
}
\usage{
github.jmcastagnetto.covid19perudata(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Peru
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
}
}

\keyword{internal}
