% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.lisphilar.covid19sir.R
\name{github.lisphilar.covid19sir}
\alias{github.lisphilar.covid19sir}
\title{Hirokazu Takaya}
\source{
https://github.com/lisphilar/covid19-sir
}
\usage{
github.lisphilar.covid19sir(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Japan
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
}
}

\keyword{internal}
