% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.m3it.covid19data.R
\name{github.m3it.covid19data}
\alias{github.m3it.covid19data}
\title{Matt Bolton}
\source{
https://github.com/M3IT/COVID-19_Data
}
\usage{
github.m3it.covid19data(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Australia
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\keyword{internal}
