% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.minciencia.datoscovid19.R
\name{github.minciencia.datoscovid19}
\alias{github.minciencia.datoscovid19}
\title{Ministerio de Ciencia, Tecnología, Conocimiento, e Innovación}
\source{
https://github.com/MinCiencia/Datos-COVID19
}
\usage{
github.minciencia.datoscovid19(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Chile
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item intensive care
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
