% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.nytimes.covid19data.R
\name{github.nytimes.covid19data}
\alias{github.nytimes.covid19data}
\title{The New York Times}
\source{
https://github.com/nytimes/covid-19-data
}
\usage{
github.nytimes.covid19data(level, fips = NULL)
}
\arguments{
\item{level}{1, 2, 3}

\item{fips}{filter by FIPS code}
}
\description{
Data source for: United States
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
}
}

\keyword{internal}
