% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.openzh.covid19.R
\name{github.openzh.covid19}
\alias{github.openzh.covid19}
\title{Specialist Unit for Open Government Data Canton of Zurich}
\source{
https://github.com/openZH/covid_19
}
\usage{
github.openzh.covid19(level, state)
}
\arguments{
\item{level}{1 (only for Liechtenstein), or 2 (only for Switzerland)}

\item{state}{one of CH (Switzerland) or FL (Liechtenstein)}
}
\description{
Data source for: Switzerland and Liechtenstein
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\keyword{internal}
