% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.oxcgrt.covidpolicytracker.R
\name{github.oxcgrt.covidpolicytracker}
\alias{github.oxcgrt.covidpolicytracker}
\title{Oxford Covid-19 Government Response Tracker}
\source{
https://github.com/OxCGRT/covid-policy-tracker
}
\usage{
github.oxcgrt.covidpolicytracker(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Worldwide
Documentation: https://covid19datahub.io/articles/doc/data.html#policy-measures
}
\keyword{internal}
