% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.ozanerturk.covid19turkeyapi.R
\name{github.ozanerturk.covid19turkeyapi}
\alias{github.ozanerturk.covid19turkeyapi}
\title{Ozan Erturk}
\source{
https://github.com/ozanerturk/covid19-turkey-api
}
\usage{
github.ozanerturk.covid19turkeyapi(level)
}
\arguments{
\item{level}{1}
}
\description{
Data source for: Turkey
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item intensive care
\item patients requiring ventilation
}
}

\keyword{internal}
