% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.pcmdpc.covid19.R
\name{github.pcmdpc.covid19}
\alias{github.pcmdpc.covid19}
\title{Ministero della Salute}
\source{
https://github.com/pcm-dpc/COVID-19
}
\usage{
github.pcmdpc.covid19(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Italy
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
}
}

\keyword{internal}
