% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ds_github.robertkochinstitut.covid19impfungenindeutschland.R
\name{github.robertkochinstitut.covid19impfungenindeutschland}
\alias{github.robertkochinstitut.covid19impfungenindeutschland}
\title{Robert Koch Institut}
\source{
https://github.com/robert-koch-institut/COVID-19-Impfungen_in_Deutschland
}
\usage{
github.robertkochinstitut.covid19impfungenindeutschland(level)
}
\arguments{
\item{level}{2, 3}
}
\description{
Data source for: Germany
}
\section{Level 2}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 3}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
