% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.swsoyee.2019ncovjapan.R
\name{github.swsoyee.2019ncovjapan}
\alias{github.swsoyee.2019ncovjapan}
\title{Su Wei}
\source{
https://github.com/swsoyee/2019-ncov-japan/blob/master/README.en.md
}
\usage{
github.swsoyee.2019ncovjapan(level, id = NULL)
}
\arguments{
\item{level}{1, 2}

\item{id}{filter by name}
}
\description{
Data source for: Japan and Cruise Ships
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\keyword{internal}
