% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.wcota.covid19br.R
\name{github.wcota.covid19br}
\alias{github.wcota.covid19br}
\title{Wesley Cota}
\source{
https://github.com/wcota/covid19br
}
\usage{
github.wcota.covid19br(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Brazil
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
}
}

\keyword{internal}
