% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.wcota.covid19br.vac.R
\name{github.wcota.covid19br.vac}
\alias{github.wcota.covid19br.vac}
\title{Wesley Cota}
\source{
https://github.com/wcota/covid19br-vac
}
\usage{
github.wcota.covid19br.vac(level)
}
\arguments{
\item{level}{3}
}
\description{
Data source for: Brazil
}
\section{Level 3}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
