% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_go.th.R
\name{go.th}
\alias{go.th}
\title{Department of Disease Control, Thailand Ministry of Public Health}
\source{
https://data.go.th/en/dataset/covid-19-daily
}
\usage{
go.th(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Thailand
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
}
}

\keyword{internal}
