% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gob.ar.R
\name{gob.ar}
\alias{gob.ar}
\title{Argentine Ministry of Health}
\source{
https://datos.gob.ar/dataset?q=covid&tags=COVID-19&sort=metadata_modified+desc
}
\usage{
gob.ar(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Argentina
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
