% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gob.pe.R
\name{gob.pe}
\alias{gob.pe}
\title{Ministerio de Salud}
\source{
https://www.datosabiertos.gob.pe/dataset/vacunaci\%C3\%B3n-contra-covid-19-ministerio-de-salud-minsa
}
\usage{
gob.pe(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Peru
}
\section{Level 1}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 2}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 3}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
