% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gouv.fr.R
\name{gouv.fr}
\alias{gouv.fr}
\title{Santé Publique France}
\source{
https://www.data.gouv.fr/fr/pages/donnees-coronavirus/
}
\usage{
gouv.fr(level = 1, reg = NULL, dep = NULL)
}
\arguments{
\item{level}{1, 2, 3}

\item{reg}{filter by region code}

\item{dep}{filter by department code}
}
\description{
Data source for: France and overseas territories
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\keyword{internal}
