% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gov.co.R
\name{gov.co}
\alias{gov.co}
\title{Instituto Nacional de Salud}
\source{
https://www.datos.gov.co/browse?category=Salud+y+Protecci\%C3\%B3n+Social&q=covid&sortBy=relevance
}
\usage{
gov.co(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Colombia
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
