% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gov.lv.R
\name{gov.lv}
\alias{gov.lv}
\title{Center for Disease Prevention and Control}
\source{
https://data.gov.lv/dati/eng/dataset?q=covid&res_format=CSV&sort=score+desc\%2C+metadata_modified+desc
}
\usage{
gov.lv(level)
}
\arguments{
\item{level}{1, 3}
}
\description{
Data source for: Latvia
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item tests
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
}
}

\keyword{internal}
