% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gov.si.R
\name{gov.si}
\alias{gov.si}
\title{Ministry of Health and National Institute for Public health}
\source{
https://www.gov.si/en/topics/coronavirus-disease-covid-19/actual-data/
}
\usage{
gov.si(level)
}
\arguments{
\item{level}{1}
}
\description{
Data source for: Slovenia
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item hospitalizations
\item intensive care
}
}

\keyword{internal}
