% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gov.tw.R
\name{gov.tw}
\alias{gov.tw}
\title{Ministry of Health and Welfare of Taiwan}
\source{
https://data.gov.tw
}
\usage{
gov.tw(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Taiwan
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item tests
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
}
}

\keyword{internal}
