% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gov.uk.R
\name{gov.uk}
\alias{gov.uk}
\title{UK Health Security Agency}
\source{
https://coronavirus.data.gov.uk
}
\usage{
gov.uk(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: United Kingdom
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item patients requiring ventilation
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item patients requiring ventilation
}
}

\keyword{internal}
