% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_GPC.R
\name{GPC}
\alias{GPC}
\title{Cruise Ship Grand Princess}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/iso_GPC.R}
}
\usage{
GPC(level)
}
\description{
Cruise Ship Grand Princess
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
confirmed cases,
deaths,
recovered.
}
}
}

\concept{Level 1}
