% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{graph_palette}
\alias{graph_palette}
\title{default palette for graph pages}
\format{
An object of class \code{character} of length 28.
}
\usage{
graph_palette
}
\description{
default palette for graph pages
}
\keyword{datasets}
