% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{growth_v_prev_calc}
\alias{growth_v_prev_calc}
\title{Calculates growth vs prevalence factors}
\usage{
growth_v_prev_calc(data, growthvar, prevvar)
}
\arguments{
\item{data}{data.frame aggregated data per region}

\item{growthvar}{growth factors}

\item{prevvar}{prevalence over 1 M}
}
\value{
factor vector
}
\description{
Calculates growth vs prevalence factors
}
\note{
factors created:
c('Low Growth and Prevalence', 'Low Growth - High Prevalence', 'High Growth - Low Prevalence', 'High Growth and Prevalence')
}
