% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gv.at.R
\name{gv.at}
\alias{gv.at}
\title{Federal Ministry of Social Affairs, Health, Care and Consumer Protection, Austria}
\source{
https://www.data.gv.at/covid-19/
}
\usage{
gv.at(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Austria
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\keyword{internal}
