% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gv_at.R
\name{gv_at}
\alias{gv_at}
\title{Federal Ministry of Social Affairs, Health, Care and Consumer Protection, Austria (BMSGPK)}
\source{
https://www.data.gv.at/covid-19/
}
\usage{
gv_at(level)
}
\description{
Imports confirmed cases, deaths, recovered, tests, hospitalizations, and vaccines at
levels 1 and 2 for Austria from BMSGPK. Confirmed cases, recovered, and deaths are
also available at level 3.
}
\keyword{internal}
