% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_healthdata.gov.R
\name{healthdata.gov}
\alias{healthdata.gov}
\title{U.S. Department of Health & Human Services}
\source{
https://healthdata.gov/browse?tags=hhs+covid-19
}
\usage{
healthdata.gov(level)
}
\arguments{
\item{level}{2}
}
\description{
Data source for: United States
}
\section{Level 2}{

\itemize{
\item tests
\item hospitalizations
\item intensive care
}
}

\keyword{internal}
