% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_healthdata.gov.R
\name{healthdata.gov}
\alias{healthdata.gov}
\title{U.S. Department of Health & Human Services}
\source{
https://healthdata.gov/browse?tags=hhs+covid-19
}
\usage{
healthdata.gov(level)
}
\description{
Imports hospitalization data and tests for United States at state level
from the U.S. Department of Health & Human Services.
}
\keyword{internal}
