% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_HRV.R
\name{HRV}
\alias{HRV}
\title{Croatia}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_HRV.R}
}
\usage{
HRV(level)
}
\description{
Croatia
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_koronavirus.hr.R}{Croatian Institute of Public Health}:
confirmed cases,
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_koronavirus.hr.R}{Croatian Institute of Public Health}:
confirmed cases,
deaths,
recovered.
}
}
}

\concept{Level 1}
\concept{Level 2}
