% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_HTI.R
\name{HTI}
\alias{HTI}
\title{Haiti}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_HTI.R}
}
\usage{
HTI(level)
}
\description{
Haiti
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
confirmed cases,
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_humdata.org.R}{Ministry of Public Health and Population of Haiti}:
confirmed cases,
deaths.
}
}
}

\concept{Level 1}
\concept{Level 2}
