% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_humdata.af.R
\name{humdata.af}
\alias{humdata.af}
\title{Afghanistan Ministry of Health}
\source{
https://data.humdata.org/dataset/afghanistan-covid-19-statistics-per-province
}
\usage{
humdata.af(level)
}
\arguments{
\item{level}{2}
}
\description{
Data source for: Afghanistan
}
\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
}
}

\keyword{internal}
