% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_humdata.org.R
\name{humdata.org}
\alias{humdata.org}
\title{Ministry of Public Health and Population of Haiti}
\source{
https://data.humdata.org/dataset/haiti-covid-19-subnational-cases
}
\usage{
humdata.org(level)
}
\arguments{
\item{level}{2}
}
\description{
Data source for: Haiti
}
\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
}
}

\keyword{internal}
