% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_impfdashboard.de.R
\name{impfdashboard.de}
\alias{impfdashboard.de}
\title{Robert Koch Institute and the Federal Ministry of Health}
\source{
https://impfdashboard.de/en/data
}
\usage{
impfdashboard.de(level)
}
\arguments{
\item{level}{1}
}
\description{
Data source for: Germany
}
\section{Level 1}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
