% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_IRL.R
\name{IRL}
\alias{IRL}
\title{Ireland}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_IRL.R}
}
\usage{
IRL(level)
}
\description{
Ireland
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_geohive.ie.R}{Health Protection Surveillance Centre (HPSC) and Health Service Executive (HSE)}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
recovered.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/IRL.csv}{Ireland Health Protection Surveillance Centre (HPSC) and Health Service Executive (HSE) (2020)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_geohive.ie.R}{Health Protection Surveillance Centre (HPSC) and Health Service Executive (HSE)}:
confirmed cases,
deaths,
recovered.
}
}
}

\concept{Level 1}
\concept{Level 2}
