% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_isciii.es.R
\name{isciii.es}
\alias{isciii.es}
\title{Centro Nacional de Epidemiología}
\source{
https://cnecovid.isciii.es/covid19/#documentaci\%C3\%B3n-y-datos
}
\usage{
isciii.es(level)
}
\arguments{
\item{level}{2, 3}
}
\description{
Data source for: Spain
}
\section{Level 2}{

\itemize{
\item confirmed cases
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item hospitalizations
\item intensive care
}
}

\keyword{internal}
