% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISO.R
\name{ISO}
\alias{ISO}
\title{THIS IS A TEMPLATE TO SUPPORT ADDITIONAL COUNTRIES}
\usage{
ISO(level, cache)
}
\description{
Copy this template in a new file
Rename the file with the ISO code of the country
Rename this function with the ISO code of the country
}
\details{
You can find ISO codes here:
https://github.com/emanuele-guidotti/COVID19/blob/master/inst/extdata/db/ISO.csv

The function must include the arguments 'level' and 'cache'. See:
https://github.com/emanuele-guidotti/COVID19/blob/master/R/covid19.R
The function cannot include additional arguments
The function must return a data.frame

Examples:
https://github.com/emanuele-guidotti/COVID19/blob/master/R/
}
\keyword{'id'.}
\keyword{Create}
\keyword{Get}
\keyword{ID}
\keyword{Merge}
\keyword{a}
\keyword{across}
\keyword{and}
\keyword{and/or}
\keyword{code}
\keyword{column}
\keyword{data}
\keyword{different}
\keyword{each}
\keyword{filter}
\keyword{filter.}
\keyword{for}
\keyword{from}
\keyword{generate}
\keyword{get}
\keyword{here...}
\keyword{here:}
\keyword{https://github.com/emanuele-guidotti/COVID19/blob/master/R/}
\keyword{if}
\keyword{internal}
\keyword{merge}
\keyword{more}
\keyword{new}
\keyword{one}
\keyword{optionally}
\keyword{or}
\keyword{package}
\keyword{return}
\keyword{row}
\keyword{source}
\keyword{sources}
\keyword{started}
\keyword{support}
\keyword{the}
\keyword{this}
\keyword{to}
\keyword{unique}
\keyword{use}
\keyword{want}
\keyword{you}
\keyword{your}
