% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iso_docstring}
\alias{iso_docstring}
\title{Generate docstring to use in the iso_ files}
\usage{
iso_docstring(ds, ...)
}
\arguments{
\item{ds}{the name of the ds_ R function}

\item{...}{arguments passed to the ds_ function}
}
\description{
Generate docstring to use in the iso_ files
}
\keyword{internal}
