% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{isoweek2date}
\alias{isoweek2date}
\title{ISO week to date}
\usage{
isoweek2date(isoweeks, day)
}
\arguments{
\item{isoweeks}{vector of isoweeks in the format YYYYMM.}

\item{day}{integer between 1 (Mondays) and 7 (Sundays).}
}
\value{
Date.
}
\description{
Converts ISO week (e.g., 202009) to date (e.g. 2020-02-24)
}
\keyword{internal}
