% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IT}
\alias{IT}
\title{Demographic indicators - Italy}
\format{\code{data.frame}}
\source{
\url{https://www.istat.it}
}
\usage{
IT
}
\description{
Tidy format dataset of Italian demographic indicators from
\href{https://www.istat.it/en/population-and-households?data-and-indicators}{Istituto Nazionale di Statistica}.
}
\seealso{
\code{\link{WB}}, \code{\link{CH}}
}
\keyword{datasets}
