% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_ITA.R
\name{ITA}
\alias{ITA}
\title{Italy}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_ITA.R}
}
\usage{
ITA(level)
}
\description{
Italy
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.pcmdpc.covid19.R}{Ministero della Salute}:
confirmed cases,
deaths,
recovered,
tests,
hospitalizations,
intensive care.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.italia.covid19opendatavaccini.R}{Commissario straordinario per l'emergenza Covid-19, Presidenza del Consiglio dei Ministri}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.pcmdpc.covid19.R}{Ministero della Salute}:
confirmed cases,
deaths,
recovered,
tests,
hospitalizations,
intensive care.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.italia.covid19opendatavaccini.R}{Commissario straordinario per l'emergenza Covid-19, Presidenza del Consiglio dei Ministri}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.pcmdpc.covid19.R}{Ministero della Salute}:
confirmed cases.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
