% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy.R
\name{italy}
\alias{italy}
\title{Italy}
\usage{
italy(type = "state", raw = FALSE)
}
\arguments{
\item{type}{one of \code{country} (data by country), \code{state} (data by region) or \code{city} (data by city). Default \code{state}, data by region.}

\item{raw}{logical. Skip data cleaning? Default \code{FALSE}.}
}
\value{
Grouped \code{tibble} (\code{data.frame}). \href{https://github.com/emanuele-guidotti/COVID19}{More info}
}
\description{
Tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.
Italian data by country, state (regions) or city.
}
\details{
See \href{https://github.com/emanuele-guidotti/COVID19}{data sources}
}
\examples{
# data by country
x <- italy("country")

# data by region
x <- italy("state")

# data by city
x <- italy("city")

}
