% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_jhucsse_git.R
\name{jhucsse_git}
\alias{jhucsse_git}
\title{Johns Hopkins Center for Systems Science and Engineering}
\source{
https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data/csse_covid_19_time_series
}
\usage{
jhucsse_git(file, cache = TRUE, level = 1, country = NULL, state = NULL)
}
\description{
Imports worldwide confirmed cases, recovered, and deaths at national level from
"Johns Hopkins Center for Systems Science and Engineering".
Confirmed cases and deaths are also available at state and county level for United States.
}
\keyword{internal}
