% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_jhunified_git.R
\name{jhunified_git}
\alias{jhunified_git}
\title{Johns Hopkins Center for Systems Science and Engineering}
\source{
https://github.com/CSSEGISandData/COVID-19_Unified-Dataset
}
\usage{
jhunified_git(iso, level)
}
\description{
Imports worldwide confirmed cases, recovered, deaths, tests, and hospitalizations at various
levels from the JHU CSSE Unified Dataset.
}
\keyword{internal}
