% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_JPN.R
\name{JPN}
\alias{JPN}
\title{Japan}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_JPN.R}
}
\usage{
JPN(level)
}
\description{
Japan
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{Japan COVID-19 Bulletin Board (2020)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.lisphilar.covid19sir.R}{Hirokazu Takaya}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/JPN.csv}{Japan COVID-19 Bulletin Board (2020)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.lisphilar.covid19sir.R}{Hirokazu Takaya}:
confirmed cases,
deaths,
recovered,
tests,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
\concept{Level 2}
