% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_koronavirus.hr.R
\name{koronavirus.hr}
\alias{koronavirus.hr}
\title{Croatian Institute of Public Health}
\source{
https://www.koronavirus.hr/otvoreni-strojno-citljivi-podaci/526
}
\usage{
koronavirus.hr(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Croatia
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
}
}

\keyword{internal}
