% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lab_percent}
\alias{lab_percent}
\title{Utility for Y labels in plots percentage}
\usage{
lab_percent(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
x labelled values
}
\description{
Utility for Y labels in plots percentage
}
