% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_calc_continent.R
\name{legend_fun}
\alias{legend_fun}
\title{Utility legend building}
\usage{
legend_fun(x, var)
}
\arguments{
\item{x}{numeric vector of map data}

\item{var}{character: variable name}
}
\value{
list legend parameters
}
\description{
Utility legend building
}
