% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_LIE.R
\name{LIE}
\alias{LIE}
\title{Liechtenstein}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/iso_LIE.R}
}
\usage{
LIE(level)
}
\description{
Liechtenstein
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/ds_admin.ch.R}{Federal Office of Public Health}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/ds_github.openzh.covid19.R}{Specialist Unit for Open Government Data Canton of Zurich}:
recovered,
patients requiring ventilation.
}
}
}

\concept{Level 1}
