% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liechtenstein.R
\name{liechtenstein}
\alias{liechtenstein}
\title{Coronavirus COVID-19 data - Liechtenstein}
\usage{
liechtenstein(raw = FALSE)
}
\arguments{
\item{raw}{logical. Skip data cleaning? Default \code{FALSE}.}
}
\value{
Return of the internal function \code{\link{covid19}}
}
\description{
Tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.
Liechtenstein data.
The data are downloaded in real-time.
}
\details{
Data pulled from \href{https://github.com/openZH/covid_19}{Open Government Data} which are communicated by official
Principality of Liechtenstein's sources.
}
\examples{
x <- liechtenstein()

}
\seealso{
\code{\link{world}}, \code{\link{diamond}}, \code{\link{italy}}, \code{\link{switzerland}}
}
