% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_countries_data_map}
\alias{load_countries_data_map}
\title{load countries  data}
\usage{
load_countries_data_map(
  destpath = system.file("./countries_data", package = "Covid19")
)
}
\arguments{
\item{destpath}{path to file}
}
\value{
countries shapefile
}
\description{
load countries  data
}
