% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_countries_datahub_map}
\alias{load_countries_datahub_map}
\title{load countries data to match with get_datahub}
\usage{
load_countries_datahub_map(
  destpath = system.file("./countries_data", package = "Covid19Mirai")
)
}
\arguments{
\item{destpath}{path to file}
}
\value{
countries shapefile
}
\description{
load countries data to match with get_datahub
}
