% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_LTU.R
\name{LTU}
\alias{LTU}
\title{Lithuania}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_LTU.R}
}
\usage{
LTU(level)
}
\description{
Lithuania
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.mpiktas.covid19lt.R}{Vaidotas Zemlys-Balevicius}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care,
patients requiring ventilation.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.mpiktas.covid19lt.R}{Vaidotas Zemlys-Balevicius}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.mpiktas.covid19lt.R}{Vaidotas Zemlys-Balevicius}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
