% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_LVA.R
\name{LVA}
\alias{LVA}
\title{Latvia}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_LVA.R}
}
\usage{
LVA(level)
}
\description{
Latvia
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gov.lv.R}{Center for Disease Prevention and Control}:
confirmed cases,
deaths,
tests.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/LVA.csv}{Oficialas statistikas portals (2019)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gov.lv.R}{Center for Disease Prevention and Control}:
confirmed cases.
}
}
}

\concept{Level 1}
\concept{Level 3}
