% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lw_positive_test_rate_calc}
\alias{lw_positive_test_rate_calc}
\title{Calculates positive test rate}
\usage{
lw_positive_test_rate_calc(conf, tests)
}
\arguments{
\item{conf}{numeric confirmed data last week}

\item{tests}{numeric tests data last week}
}
\value{
x numeric positive tests rate
}
\description{
Calculates positive test rate
}
