% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_model_cml_linear}
\alias{make_model_cml_linear}
\title{Split dataset in a list of data frames
Create a cumulative sum
Filter to a time frame defined by time_unit_extent
Model data for each country using a linear regression of the cumulative count}
\usage{
make_model_cml_linear(
  series,
  df = dfc_ecdc,
  time_unit_extent = 12,
  min_sum = 100
)
}
\description{
Split dataset in a list of data frames
Create a cumulative sum
Filter to a time frame defined by time_unit_extent
Model data for each country using a linear regression of the cumulative count
}
