% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_model_cnt_linear}
\alias{make_model_cnt_linear}
\title{Split dataset into a list of time-series, one for each country (splitting uses the iso_a3 variable)
Filter the time-series to a time-frame defined by time_unit_extent
Fill the gaps in the time-series, assuming that at missing date, cases and deaths = 0 
Smoothing using moving average with a parameterable time-window (in days)
Model time-series using a linear regression}
\usage{
make_model_cnt_linear(series, df = df_ecdc, time_unit_extent = 12,
  ma_window = 3, min_sum = 30)
}
\description{
Split dataset into a list of time-series, one for each country (splitting uses the iso_a3 variable)
Filter the time-series to a time-frame defined by time_unit_extent
Fill the gaps in the time-series, assuming that at missing date, cases and deaths = 0 
Smoothing using moving average with a parameterable time-window (in days)
Model time-series using a linear regression
}
