% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_data}
\alias{map_data}
\title{Map Data}
\usage{
map_data(x, map)
}
\arguments{
\item{x}{\code{data.frame}}

\item{map}{named vector. Map columns of \code{x} and subset.}
}
\value{
\code{data.frame}
}
\description{
Subset a \code{data.frame} and change column names.
}
\examples{
\dontrun{

x <- mtcars

map_data(x, c(
'cyl' = 'Cylinders',
'hp'  = 'Gross horsepower'
))

}

}
