% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_calc_continent.R
\name{map_popup_data}
\alias{map_popup_data}
\title{Utility for popup message in map}
\usage{
map_popup_data(data, nam, ind, namvar, textvar)
}
\arguments{
\item{data}{map data}

\item{nam}{character: component of country names from data, NAME}

\item{ind}{character: component of values from data, indicator}

\item{namvar}{character: vector, additional variable names}

\item{textvar}{character: vector, textt for the additional variables}
}
\value{
vector pop up messages, html
}
\description{
Utility for popup message in map
}
