% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{map_values}
\alias{map_values}
\title{Map values}
\usage{
map_values(x, map, force = FALSE)
}
\arguments{
\item{x}{vector.}

\item{map}{named vector mapping names to values.}
}
\value{
\code{vector}.
}
\description{
Map values of a vector.
}
\keyword{internal}
