% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{merge}
\alias{merge}
\title{Merge Two Data Frames}
\usage{
merge(..., fill = FALSE)
}
\arguments{
\item{...}{arguments passed to \code{\link[base]{merge}}}

\item{fill}{whether to merge non-NA values of duplicated columns}
}
\value{
return value of \code{\link[base]{merge}}
}
\description{
Merge two data frames by common columns or row names, or do other versions of database join operations.
Drop common columns of the second data.frame.
}
\keyword{internal}
