% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{merge_pop_data}
\alias{merge_pop_data}
\title{merge data}
\usage{
merge_pop_data(data, popdata)
}
\arguments{
\item{data}{data.frame}

\item{popdata}{data.frame result of get_pop_data}
}
\value{
global tibble of confirmed, deaths, active and recovered for each day by population
}
\description{
merge data
}
\examples{
\dontrun{
orig_data <- get_timeseries_full_data() \%>\%
              get_timeseries_by_contagion_day_data()
data <- orig_data \%>\% align_country_names()

popdata <- get_pop_data()

}
}
