% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_conf_case}
\alias{message_conf_case}
\title{Message text, selection of confirmed cases}
\usage{
message_conf_case(where = "Countries", ncases = 1000, suffix = "can be chosen")
}
\arguments{
\item{where}{character string, where text happens}

\item{ncases}{numeric string, minimum number of confirmed cases}

\item{suffix}{character string, end of text}
}
\value{
character vector
}
\description{
Message text, selection of confirmed cases
}
