% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_firstday}
\alias{message_firstday}
\title{Message text, selection of confirmed cases}
\usage{
message_firstday(ncases, var = "confirmed")
}
\arguments{
\item{ncases}{numeric string, minimum number of confirmed cases}

\item{var}{character string, variable used}
}
\value{
character vector
}
\description{
Message text, selection of confirmed cases
}
