% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_hosp_data}
\alias{message_hosp_data}
\title{Message text, missing data}
\usage{
message_hosp_data(
  what = "Hospitalised, Vaccinated and Test",
  where = "some countries and areas",
  suffix = "where available"
)
}
\arguments{
\item{what}{character string, text for missing variables}

\item{where}{character string, where text happens}

\item{suffix}{character string, end of text}
}
\value{
character vector
}
\description{
Message text, missing data
}
