% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_missing_country_days}
\alias{message_missing_country_days}
\title{Message text, missing days for given countries in the data}
\usage{
message_missing_country_days(data)
}
\arguments{
\item{country_data}{list, names are the number of days, the components are the country names}
}
\value{
character vector with message
}
\description{
Message text, missing days for given countries in the data
}
