% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_missing_data}
\alias{message_missing_data}
\title{Message text, missing data}
\usage{
message_missing_data(
  what = "Recovered, Hospitalized and Tests",
  where = "some countries and areas"
)
}
\arguments{
\item{what}{character string, text for missing variables}

\item{where}{character string, where text happens}
}
\value{
character vector
}
\description{
Message text, missing data
}
