% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_missing_recovered}
\alias{message_missing_recovered}
\title{Message text, recovered}
\usage{
message_missing_recovered(
  what = "Recovered",
  where = "Some countries and areas"
)
}
\arguments{
\item{what}{character string, text for missing variables}

\item{where}{character string, where text happens}
}
\value{
character vector
}
\description{
Message text, recovered
}
