% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message_subcountries}
\alias{message_subcountries}
\title{creates message of countries within subcountries}
\usage{
message_subcountries(data, area, region)
}
\arguments{
\item{data}{data.frame aggregated data per region}

\item{area}{character main area}

\item{region}{character sub area}
}
\value{
list messages printing Country.Region within subcontinent
}
\description{
creates message of countries within subcountries
}
