% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_caseBoxes.R
\name{mod_caseBoxes_server}
\alias{mod_caseBoxes_server}
\title{caseBoxes Server Function}
\usage{
mod_caseBoxes_server(
  input,
  output,
  session,
  counts,
  hosp = FALSE,
  vax = NULL,
  renderui = FALSE
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{counts}{Reactive expression yielding the named vector of cases by type.}

\item{hosp}{logical, if TRUE hospitalised box variables is added.}

\item{vax}{character, output id for the vaccination box}

\item{renderui}{logical, if TRUE shiny::renderUI is used}
}
\description{
caseBoxes Server Function
}
