% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_caseBoxes.R
\name{mod_caseBoxes_ui}
\alias{mod_caseBoxes_ui}
\title{caseBoxes UI Function}
\usage{
mod_caseBoxes_ui(id, hosp = FALSE, outputui = FALSE)
}
\arguments{
\item{id, }{Internal parameters for {shiny}.}

\item{hosp}{logical, if TRUE hospitalised box variables is added.}

\item{outputui}{logical, if TRUE shiny::uiOutput is used}
}
\description{
A shiny Module for displaying the number of cases by type as
  colored boxes.
}
\keyword{internal}
