% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_example_module.R
\name{mod_example_module_ui}
\alias{mod_example_module_ui}
\alias{mod_example_module_server}
\title{mod_example_module_ui and mod_example_module_server}
\usage{
mod_example_module_ui(id)

mod_example_module_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
