% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_leaflet.R
\name{mod_leaflet_ui}
\alias{mod_leaflet_ui}
\alias{mod_leaflet_server}
\title{mod_leaflet_ui and mod_leaflet_server}
\usage{
mod_leaflet_ui(id)

mod_leaflet_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{plot_years}{Years to plot}
}
\description{
A shiny Module.
}
\keyword{internal}
