% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_no_data_server.R
\name{mod_nodata_areaplot_server}
\alias{mod_nodata_areaplot_server}
\title{No data in variable text Server Function}
\usage{
mod_nodata_areaplot_server(input, output, session, country, what = "Vaccines")
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}

\item{country}{character country name}

\item{what}{variable name}
}
\description{
No data in variable text Server Function
}
