% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vaccines_text.R
\name{mod_vaccines_text_server}
\alias{mod_vaccines_text_server}
\title{Vaccines text Server Function}
\usage{
mod_vaccines_text_server(input, output, session, df, dftoday)
}
\arguments{
\item{df}{dataset whole time series.}

\item{dftoday}{dataset with today and weekly data.}
}
\description{
Vaccines text Server Function
}
