% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vaccines_text.R
\name{mod_vaccines_text_ui}
\alias{mod_vaccines_text_ui}
\title{caseBoxes UI Function}
\usage{
mod_vaccines_text_ui(id)
}
\arguments{
\item{id, }{Internal parameters for {shiny}.}
}
\description{
A shiny Module for displaying stats about vaccination
}
