% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_mzcr.cz.R
\name{mzcr.cz}
\alias{mzcr.cz}
\title{Ministry of Health of the Czech Republic}
\source{
https://onemocneni-aktualne.mzcr.cz/api/v2/covid-19
}
\usage{
mzcr.cz(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Czech Republic
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
}
}

\keyword{internal}
