% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{new_total_colors}
\alias{new_total_colors}
\title{Color Palette}
\format{
An object of class \code{character} of length 2.
}
\usage{
new_total_colors
}
\description{
Color Palette
}
\keyword{datasets}
