% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_NLD.R
\name{NLD}
\alias{NLD}
\title{Netherlands}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_NLD.R}
}
\usage{
NLD(level)
}
\description{
Netherlands
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_rivm.nl.R}{National Institute for Public Health and the Environment}:
confirmed cases,
deaths.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_rivm.nl.R}{National Institute for Public Health and the Environment}:
confirmed cases,
deaths.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_rivm.nl.R}{National Institute for Public Health and the Environment}:
confirmed cases,
deaths.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
