% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_nytimes_git.R
\name{nytimes_git}
\alias{nytimes_git}
\title{The New York Times}
\source{
https://github.com/nytimes/covid-19-data
}
\usage{
nytimes_git(level, cache = FALSE, fips = NULL)
}
\description{
Imports confirmed cases and deaths for United States at national, state, and county level
from The New York Times.
}
\keyword{internal}
