% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{oneM_pop_calc}
\alias{oneM_pop_calc}
\title{Calculates ratio over 1M pop}
\usage{
oneM_pop_calc(x, pop)
}
\arguments{
\item{x}{numeric data}

\item{pop}{numeric vector population size}
}
\value{
x numeric x rescaled by pop size
}
\description{
Calculates ratio over 1M pop
}
