% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_ourworldindata.org.R
\name{ourworldindata.org}
\alias{ourworldindata.org}
\title{Our World in Data}
\source{
https://covid.ourworldindata.org
}
\usage{
ourworldindata.org(level = 1, id = NULL)
}
\arguments{
\item{level}{1, 2}

\item{id}{filter by ISO code if level=1 or by name of state if level=2}
}
\description{
Data source for: Worldwide
}
\section{Level 1}{

\itemize{
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
