% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_calc_continent.R
\name{pal_fun}
\alias{pal_fun}
\title{Utility derive palette given data}
\usage{
pal_fun(var, x)
}
\arguments{
\item{var}{character: variable name}

\item{x}{numeric vector of map data}
}
\value{
palette
}
\description{
Utility derive palette given data
}
