% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.pcmdpc.covid19.R
\name{github.pcmdpc.covid19}
\alias{github.pcmdpc.covid19}
\title{Ministero della Salute}
\source{
https://github.com/pcm-dpc/COVID-19
}
\usage{
github.pcmdpc.covid19(level)
}
\description{
Imports confirmed cases, deaths, recovered, tests, hospitalizations and
intensive care from the Italian Ministry of Health. The data are available
at national and regional level. Only confirmed cases are available at province
level.
}
\keyword{internal}
