% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_PER.R
\name{PER}
\alias{PER}
\title{Peru}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/iso_PER.R}
}
\usage{
PER(level)
}
\description{
Peru
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.jmcastagnetto.covid19perudata.R}{Jesus M. Castagnetto}:
confirmed cases,
deaths,
recovered,
tests,
hospitalizations,
intensive care,
patients requiring ventilation.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_gob.pe.R}{Ministerio de Salud}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.jmcastagnetto.covid19perudata.R}{Jesus M. Castagnetto}:
confirmed cases,
deaths,
recovered,
tests.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_gob.pe.R}{Ministerio de Salud}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}
}

\concept{Level 1}
\concept{Level 2}
