% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all_highlight}
\alias{plot_all_highlight}
\title{plot all countries but highlight first 10}
\usage{
plot_all_highlight(
  df,
  log = FALSE,
  text = "",
  percent = FALSE,
  date_x = FALSE,
  g_palette = graph_palette
)
}
\arguments{
\item{df}{data.frame with column called Date and Value column to plot}

\item{log}{logical for applying log scale}

\item{text}{element for tooltip}

\item{percent}{logical to make the y axis in percent}

\item{date_x}{logical to convert x-axis labels to dates}

\item{g_palette}{character vector of colors for the graph and legend}
}
\description{
plot all countries but highlight first 10
}
