% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_plots.R
\name{plot_all_highlight_10}
\alias{plot_all_highlight_10}
\title{plot all countries but highlight first 10}
\usage{
plot_all_highlight_10(df, log = F, text = "")
}
\arguments{
\item{df}{data.frame with column called Date and Value column to plot}

\item{log}{logical for applying log scale}

\item{text}{element for tooltip}
}
\description{
plot all countries but highlight first 10
}
