% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_data.R
\name{plot_country_data}
\alias{plot_country_data}
\title{plot_country_data}
\usage{
plot_country_data(
  start_date = "2-22-2020",
  end_date = "10-27-2020",
  country1 = "Iran",
  type1 = "confirmed"
)
}
\arguments{
\item{start_date}{the first day of the interval}

\item{country1}{the last country that you want to see its data.}

\item{type1}{type of data that you want to see it can be 'confirmed' 'deaths' and 'recovered'}

\item{last_date}{the last day of the interval}
}
\value{
A plot
}
\description{
plots corona data of a country in a given interval
}
\examples{
plot_country_data(start_date = '2-22-2020', end_date = '10-27-2020', country1 = 'Iran', type1='confirmed')
}
