% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_day_zero.R
\name{plot_day_zero}
\alias{plot_day_zero}
\title{Plot day zero}
\usage{
plot_day_zero(
  countries = c("Italy", "Spain", "France", "US", "Germany"),
  ylog = TRUE,
  day0 = 150,
  cumulative = TRUE,
  time_before = 0,
  add_markers = FALSE,
  line_size = 1.5,
  max_date = Sys.Date(),
  calendar = FALSE
)
}
\arguments{
\item{countries}{Character vector of country names}

\item{ylog}{Whether the y-axis should be on log scale}

\item{day0}{An integer, the number of cumulative cases at which the outbreak is considered to have started}

\item{cumulative}{Whether to count cases cumulatively}

\item{time_before}{How many days before outbreak to show}

\item{line_size}{Size of line}

\item{max_date}{The maximum date}

\item{calendar}{Whether to plot by calendar date}

\item{add_markets}{Whether to show lines / circle at outbreak start}
}
\description{
Generate a plot with time adjusted for the day at which the outbreak is considered to have started
}
