% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_day_zero.R
\name{plot_day_zero_esp}
\alias{plot_day_zero_esp}
\title{Plot day zero for Spain}
\usage{
plot_day_zero_esp(
  ccaa = c("Cataluña", "Madrid"),
  deaths = FALSE,
  ylog = TRUE,
  day0 = 150,
  cumulative = TRUE,
  time_before = 0,
  add_markers = FALSE,
  line_size = 1.5,
  max_date = Sys.Date(),
  pop = FALSE
)
}
\arguments{
\item{ccaa}{Character vector of ccaa names}

\item{deaths}{Whether to show deaths rather than cases}

\item{ylog}{Whether the y-axis should be on log scale}

\item{day0}{An integer, the number of cumulative cases at which the outbreak is considered to have started}

\item{cumulative}{Whether to count cases cumulatively}

\item{time_before}{How many days before outbreak to show}

\item{line_size}{Size of line}

\item{max_date}{The maximum date}

\item{pop}{Adjust for population}

\item{add_markets}{Whether to show lines / circle at outbreak start}
}
\description{
Generate a plot with time adjusted for the day at which the outbreak is considered to have started
}
