% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot interactive network}
\usage{
plot_network(n = 100, contacts = 1, sd = 1, save = NULL, sick = 1)
}
\arguments{
\item{n}{Number of units (families, for example)}

\item{contacts}{Average number of contacts between families}

\item{sd}{Standard deviation in number of contacts}

\item{sick}{Percent sick}
}
\description{
Generate a plot of network connections
}
