% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rate_hist}
\alias{plot_rate_hist}
\title{plot rate as hist}
\usage{
plot_rate_hist(df, color, percent = F, y_min = 0)
}
\arguments{
\item{df}{data.frame}

\item{color}{string used to define color}

\item{percent}{logical to make the y axis in percent}

\item{y_min}{min value on y axis}
}
\value{
ggplot plot
}
\description{
plot rate as hist
}
