% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_world_map.R
\name{plot_world_map}
\alias{plot_world_map}
\title{plot_world_map}
\usage{
plot_world_map(day1 = 27, month1 = 10, year1 = 20, type1 = "confirmed")
}
\arguments{
\item{day1}{the day that you want to see its data}

\item{month1}{the month that you want to see its data}

\item{type1}{the type of data that you want to see it can be 'confirmed' 'deaths' and 'recovered'}

\item{year}{the year that you want to see its data}
}
\value{
A plot
}
\description{
shows corona data of all countries on the world world map on the desired date.
}
\examples{
plot_world_map(day1 = 27,month1=10,year1=20,type1='confirmed')
}
