% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDeltaDetected}
\alias{plotDeltaDetected}
\title{plotDeltaDetected}
\usage{
plotDeltaDetected(covid19Filtered, plotPath, plotType = "smooth",
  plotCurve = c("detected", "deceased"), ylims = NA, ncols = 4)
}
\arguments{
\item{covid19Filtered}{data.frame with (filtered) data}

\item{plotPath}{full path for storing the plot}

\item{plotType}{type of plot to use. Options: c("line", "smooth")}

\item{plotCurve}{Curve(s) to be plot. Options: c("detected", "deceased", "active)}

\item{ylims}{vector with y-coordinate plot limits}

\item{ncols}{number of columns to represent plot facets}
}
\description{
plots the (daily) increase of cases
}
