% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotNewCasesVsTotalCases}
\alias{plotNewCasesVsTotalCases}
\title{plotNewCasesVsTotalCases}
\usage{
plotNewCasesVsTotalCases(covid19Filtered, plotPath, plotType = "smooth",
  plotCurve = "detected", xlims = NULL, ylims = NULL)
}
\arguments{
\item{covid19Filtered}{data.frame with (filtered) data}

\item{plotPath}{full path for storing the plot}

\item{plotType}{type of plot to use. Options: c("line", "smooth")}

\item{plotCurve}{Curve to be plot (only one curve per plot!).
Options: c("detected", "deceased", "active)}

\item{xlims}{vector with x-coordinate plot limits}

\item{ylims}{vector with y-coordinate plot limits}
}
\description{
plots new cases (daily) vs total cases in log scale
}
