% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTimeToDouble}
\alias{plotTimeToDouble}
\title{plotTimeToDouble}
\usage{
plotTimeToDouble(covid19Filtered, plotPath, plotType = "smooth",
  ylims = c(0, 20), ncols = 4)
}
\arguments{
\item{covid19Filtered}{data.frame with (filtered) data}

\item{plotPath}{full path for storing the plot}

\item{plotType}{type of plot to use. Options: c("line", "smooth")}

\item{ylims}{vector with y-coordinate plot limits}

\item{ncols}{number of columns to represent plot facets}
}
\description{
plots the time to double
}
