% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTotalCases}
\alias{plotTotalCases}
\title{plotTotalCases}
\usage{
plotTotalCases(covid19Filtered, plotPath, ylims = NA, ncols = 4)
}
\arguments{
\item{covid19Filtered}{data.frame with (filtered) data}

\item{plotPath}{full path for storing the plot}

\item{ylims}{vector with y-coordinate plot limits}

\item{ncols}{number of columns to represent plot facets}
}
\description{
plots the total number of cases
}
