% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prefix_colors}
\alias{prefix_colors}
\title{Color Palette for new and last week variables}
\usage{
prefix_colors(cc = .case_colors, prefix = "new")
}
\arguments{
\item{cc}{vector \code{.case_colors}}

\item{prefix}{character, new or lw}
}
\description{
Color Palette for new and last week variables
}
