% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prefix_var}
\alias{prefix_var}
\title{Add prefix new and last week variables}
\usage{
prefix_var(cc = .case_colors, prefix = c("lw", "new"))
}
\arguments{
\item{cc}{vector \code{.case_colors}}

\item{prefix}{character, new or lw or both}
}
\description{
Add prefix new and last week variables
}
