% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_day_zero.R
\name{prepare_day_zero_data}
\alias{prepare_day_zero_data}
\title{Prepare day zero data}
\usage{
prepare_day_zero_data(
  countries = c("Italy", "Spain", "US", "Germany"),
  day0 = 150,
  cumulative = TRUE,
  time_before = 0,
  max_date = Sys.Date(),
  deaths = FALSE,
  pop = FALSE,
  pop_adjustor = 1e+06,
  by_district = FALSE,
  districts = NULL,
  roll = 0,
  roll_fun = "mean"
)
}
\arguments{
\item{countries}{Character vector of country names}

\item{day0}{An integer, the number of cumulative cases at which the outbreak is considered to have started}

\item{cumulative}{Whether to count cases cumulatively}

\item{time_before}{How many days before outbreak to show}

\item{max_date}{The maximum date}

\item{deaths}{Whether to use deaths instead of cases}

\item{pop}{Adjust by population}

\item{pop_adjustor}{Deaths per x (default million) when doing population adjustment}

\item{by_district}{Show by district}

\item{districts}{Show by districts, default FALSE}

\item{roll}{How many days to get rolling values for}

\item{roll_fun}{Mean or Sum}
}
\description{
Generate a plot with time adjusted for the day at which the outbreak is considered to have started
}
