% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_PRT.R
\name{PRT}
\alias{PRT}
\title{Portugal}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/iso_PRT.R}
}
\usage{
PRT(level)
}
\description{
Portugal
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.dssgpt.covid19ptdata.R}{Data Science for Social Good Portugal}:
confirmed cases,
deaths,
recovered,
hospitalizations,
intensive care.
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.dssgpt.covid19ptdata.R}{Data Science for Social Good Portugal}:
confirmed cases,
deaths.
}
}
}

\concept{Level 1}
\concept{Level 2}
