% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{rate_vars}
\alias{rate_vars}
\title{define character vector of rate variables for labels with %}
\format{
An object of class \code{character} of length 1.
}
\usage{
rate_vars
}
\description{
define character vector of rate variables for labels with %
}
\keyword{datasets}
