% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.excel}
\alias{read.excel}
\title{Data Input (excel)}
\usage{
read.excel(path, cache = FALSE, sheet = NA, ...)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{cache}{logical. Memory caching? Default \code{FALSE}.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet). Ignored if the sheet is specified via range. If neither argument specifies the sheet, defaults to all sheets.}

\item{...}{arguments passed to \code{\link[readxl]{read_excel}}}
}
\value{
list of \code{data.frame}
}
\description{
Read xls and xlsx files.
}
\examples{
\dontrun{

url <- "https://epistat.sciensano.be/Data/COVID19BE.xlsx"
x   <- read.excel(url, cache = TRUE)  

}

}
