% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{read.xsv}
\alias{read.xsv}
\title{Data Input (csv)}
\usage{
read.xsv(file, select, ...)
}
\arguments{
\item{file}{the path to the csv file}

\item{select}{character vector of column names to extract}

\item{...}{additional arguments passed to \code{\link[data.table]{fread}}}
}
\value{
data.table
}
\description{
Reads specific columns of a csv file by using the xsv command line utility
https://github.com/BurntSushi/xsv
}
\keyword{internal}
