% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{read.zip}
\alias{read.zip}
\title{Data Input (zip)}
\usage{
read.zip(zip, files, cache = FALSE, ...)
}
\arguments{
\item{zip}{path (url) to the zip folder.}

\item{files}{vector of filenames to read inside the zip folder.}

\item{cache}{logical. Memory caching? Default \code{FALSE}.}

\item{...}{arguments passed to \code{\link{read.csv}} or \code{\link{read.excel}}.}
}
\value{
list of \code{data.frames}
}
\description{
Reads files from a zip folder.
}
\keyword{internal}
