% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_ecdc.R
\name{read_cases_ecdc}
\alias{read_cases_ecdc}
\title{Reads COVID-19 cases and deaths from the ECDC}
\usage{
read_cases_ecdc()
}
\value{
\code{tbl} with 11 columns:
\itemize{
\item \code{date} : (Date)
\item \code{dateRep} : (character)
\item \code{day} : (integer)
\item \code{month} : (integer)
\item \code{year} : (integer)
\item \code{cases} : (integer)
\item \code{deaths} : (integer) number of new deaths that occurred on the given day.
This is \emph{not} the cumulative total of deaths up to that day.
\item \code{countriesAndTerritories} : (character)
\item \code{geoId} : (character) 2-character ISO-3166 country code.
\item \code{countryterritoryCode} : (character) 3-character ISO-3166 country code.
\item \code{popData2018} : (integer)
}
}
\description{
Reads a \href{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}{dataset} with COVID-19 cases and deaths, which is maintained by the European Centre for Disease Prevention and Control (ECDC).
}
\details{
\itemize{
\item Reads an \href{https://www.ecdc.europa.eu/}{ECDC} \href{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}{dataset} of  \href{https://www.who.int/emergencies/diseases/novel-coronavirus-2019}{COVID-19} cases and deaths.
\item Adds a \code{date} column of class \code{Date}, derived from the \code{dateRep} column, which is a string.
}
}
