% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_confirmed_cases}
\alias{read_confirmed_cases}
\title{Reads cumulative total of COVID-19 cases.}
\usage{
read_confirmed_cases()
}
\value{
tbl with 6 columns: ("country_region", "province_state", "lat", "long", "date", "cumulative_total")
}
\description{
Reads cumulative total of COVID-19 confirmed cases by day from a public dataset published by Johns Hopkins University (JHU).
}
\details{
\itemize{
\item Reads a \href{https://github.com/CSSEGISandData/COVID-19/blob/master/csse_covid_19_data/csse_covid_19_time_series/time_series_19-covid-Confirmed.csv}{JHU CSSE data file} of confirmed \href{https://www.who.int/emergencies/diseases/novel-coronavirus-2019}{COVID-19} cases.
\item Reformats data as a \href{https://r4ds.had.co.nz/tidy-data.html}{tidy dataset}.
\item Reformats the column names to \href{https://en.wikipedia.org/wiki/Snake_case}{snake case}.
}
}
