% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_jhu_csse.R
\name{read_deaths_jhu_csse}
\alias{read_deaths_jhu_csse}
\title{Reads cumulative total of COVID-19 deaths.}
\usage{
read_deaths_jhu_csse()
}
\value{
\code{tbl} with 6 columns: \code{country_region}, \code{province_state}, \code{lat}, \code{long}, \code{date}, \code{cumulative_total}. Note: The country data for the United States and most other countries is \emph{not} sub-divided by \code{province_state}, so that column has many \code{NA} values.
}
\description{
Reads cumulative total of COVID-19 deaths by day from a public dataset published by Johns Hopkins University (JHU).
}
\details{
\itemize{
\item Reads a \href{https://github.com/CSSEGISandData/COVID-19}{JHU CSSE} data file, \href{https://github.com/CSSEGISandData/COVID-19/blob/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_global.csv}{time_series_19-covid-Deaths.csv}, of deaths from the \href{https://www.who.int/emergencies/diseases/novel-coronavirus-2019}{COVID-19} disease.
\item Reformats data as a \href{https://r4ds.had.co.nz/tidy-data.html}{tidy dataset}.
\item Reformats the column names to \href{https://en.wikipedia.org/wiki/Snake_case}{snake case}.
}
}
