% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_nytimes.R
\name{read_state_cases_nytimes}
\alias{read_state_cases_nytimes}
\title{Reads COVID-19 cases and deaths by state from the NY Times}
\usage{
read_state_cases_nytimes()
}
\value{
\code{tbl} with 5 columns:
\itemize{
\item \code{date}
\item \code{state}
\item \code{fips}
\item \code{cases}
\item \code{deaths}
}
}
\description{
\itemize{
\item Reads a \href{https://www.nytimes.com/interactive/2020/us/coronavirus-us-cases.html}{NY Times} \href{https://github.com/nytimes/covid-19-data/blob/master/us-states.csv}{dataset} of  \href{https://www.who.int/emergencies/diseases/novel-coronavirus-2019}{COVID-19} cases and deaths by state.
}
}
