% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state_populations.R
\name{read_state_populations}
\alias{read_state_populations}
\title{Reads state population estimates from the United States Census Bureau}
\usage{
read_state_populations()
}
\value{
\code{tbl} with 3 columns:
\itemize{
\item \code{state_name}
\item \code{year}
\item \code{population_estimate}
}
}
\description{
\itemize{
\item Reads a United States Census Bureau \href{https://www.census.gov/newsroom/press-kits/2019/national-state-estimates.html}{dataset} of estimated state populations for 2019.
}
}
