% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{readDataCovid}
\alias{readDataCovid}
\title{readDataCovid}
\usage{
readDataCovid(dataFile, col_types = "ciiic")
}
\arguments{
\item{dataFile}{full path to data file}

\item{col_types}{types of columns for reading the table}
}
\value{
data frame with the data
}
\description{
reads data from Covid19 in tsv format. Required columns: country, detected, healed, deceased, date.
}
