% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repo}
\alias{repo}
\title{Generate link to the GitHub repository}
\usage{
repo(x)
}
\arguments{
\item{x}{name of the iso_ or ds_ function}
}
\description{
Generate link to the GitHub repository
}
\keyword{internal}
