% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{rescale_df_contagion}
\alias{rescale_df_contagion}
\title{Rescale contagion day for countries with at least n cases and outbreaks longer than w days}
\usage{
rescale_df_contagion(df, n, w, group = "Country.Region")
}
\arguments{
\item{df}{data.frame}

\item{n}{number of cases}

\item{w}{days of outbreak}

\item{group}{character Country.Region or continent or subcontinent}
}
\description{
Rescale contagion day for countries with at least n cases and outbreaks longer than w days
}
