% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_rivm.nl.R
\name{rivm.nl}
\alias{rivm.nl}
\title{National Institute for Public Health and the Environment}
\source{
https://data.rivm.nl/covid-19/
}
\usage{
rivm.nl(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Netherlands
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item deaths
}
}

\keyword{internal}
