% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rollAvg}
\alias{rollAvg}
\title{compute rolling Weekly average}
\usage{
rollAvg(x, date, days = 7)
}
\arguments{
\item{x}{numeric vector of values}

\item{date}{Date vector of dates}

\item{days}{number of days over which average to be computed}
}
\value{
x as partial rolling average
}
\description{
compute rolling Weekly average
}
