% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_calc_continent.R
\name{roundlab}
\alias{roundlab}
\title{Utility round labels in domain and map}
\usage{
roundlab(y)
}
\arguments{
\item{x}{numeric vector of map data}
}
\value{
numeric rounded x value
}
\description{
Utility round labels in domain and map
}
