% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundUp}
\alias{roundUp}
\title{Round up to the next decine
Ref: https://stackoverflow.com/questions/6461209/how-to-round-up-to-the-nearest-10-or-100-or-x}
\usage{
roundUp(x)
}
\arguments{
\item{x}{number to round}
}
\value{
rounded up number
}
\description{
Round up to the next decine
Ref: https://stackoverflow.com/questions/6461209/how-to-round-up-to-the-nearest-10-or-100-or-x
}
