% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_sciensano.be.R
\name{sciensano.be}
\alias{sciensano.be}
\title{Sciensano, the Belgian Institute for Health}
\source{
https://epistat.wiv-isp.be/covid/
}
\usage{
sciensano.be(level)
}
\arguments{
\item{level}{1, 2, 3}
}
\description{
Data source for: Belgium
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item tests
\item total vaccine doses administered
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
\item intensive care
\item patients requiring ventilation
}
}

\keyword{internal}
