% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_sciensano_be.R
\name{sciensano_be}
\alias{sciensano_be}
\title{Sciensano, the Belgian institute for health}
\source{
https://epistat.wiv-isp.be/covid/
}
\usage{
sciensano_be(level)
}
\description{
Imports confirmed cases, hospitalizations, deaths, tests, and vaccines for Belgium
at national, regional, and province level from Sciensano. The number of deaths is not
available at the province level.
}
\details{
Sciensano uses the following codes.
\itemize{
\item For vaccines requiring 2 doses: A for first dose, B for second dose;
\item C for vaccine requiring only 1 dose
\item E for extra dose of vaccine administered since the 9th of September 2021
See https://epistat.sciensano.be/COVID19BE_codebook.pdf
}

We use A+C to compute vaccines_1 (at least one dose) and B+C to compute vaccines_2 (fully vaccinated).
}
\keyword{internal}
