% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secondline_theme}
\alias{secondline_theme}
\alias{noaxislab_theme}
\title{Additional Theme for second line in plot}
\usage{
secondline_theme()

noaxislab_theme()
}
\description{
Additional Theme for second line in plot

Remove axis labels
}
