% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_group_plots_server.R
\name{section_info}
\alias{section_info}
\title{info text on hospitalized}
\usage{
section_info(type, infotext)
}
\arguments{
\item{type}{character, .}

\item{infotext}{logical, print section info}
}
\description{
info text on hospitalized
}
