% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_group_plots_server.R
\name{section_title}
\alias{section_title}
\title{info text on hospitalized}
\usage{
section_title(type, tit)
}
\arguments{
\item{type}{character, .}

\item{tit}{logical, print title}
}
\description{
info text on hospitalized
}
