% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{select_countries_n_cases_w_days}
\alias{select_countries_n_cases_w_days}
\title{Select only countries which had at least n cases and outbreaks longer than w days}
\usage{
select_countries_n_cases_w_days(df, n, w)
}
\arguments{
\item{df}{data.frame}

\item{n}{number of cases}

\item{w}{days of outbreak}
}
\description{
Select only countries which had at least n cases and outbreaks longer than w days
}
