% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_SHN.R
\name{SHN}
\alias{SHN}
\title{Saint Helena, Ascension and Tristan da Cunha}
\source{
https://github.com/covid19datahub/COVID19/blob/10eca84e95eb6e62d36b801bfb08b59de7e6c080/R/iso_SHN.R
}
\usage{
SHN(level, ...)
}
\description{
Data available at level 1 (nation).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/10eca84e95eb6e62d36b801bfb08b59de7e6c080/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}
(confirmed cases, recovered, deaths);
\href{https://github.com/covid19datahub/COVID19/blob/10eca84e95eb6e62d36b801bfb08b59de7e6c080/R/ds_ourworldindata.org.R}{Our World in Data}
(tests, hospitalizations, vaccines);
\href{https://data.worldbank.org/indicator/SP.POP.TOTL}{World Bank Open Data}
(population 2018).
}

\concept{level 1}
