% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{simple_cap}
\alias{simple_cap}
\title{Simple Cap Function}
\usage{
simple_cap(s, strict = FALSE)
}
\arguments{
\item{s}{is the string to parse.}

\item{strict}{is a flag that allows to parse an entire word.}
}
\description{
Simple Cap Function
}
