% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_ssi.dk.R
\name{ssi.dk}
\alias{ssi.dk}
\title{Statens Serum Institut}
\source{
https://covid19.ssi.dk/overvagningsdata
}
\usage{
ssi.dk(level)
}
\arguments{
\item{level}{2, 3}
}
\description{
Data source for: Denmark
}
\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item tests
\item people with at least one vaccine dose
\item people fully vaccinated
\item hospitalizations
}
}

\section{Level 3}{

\itemize{
\item confirmed cases
\item tests
\item people with at least one vaccine dose
\item people fully vaccinated
}
}

\keyword{internal}
