% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_group_plots_server.R
\name{stringency_info}
\alias{stringency_info}
\title{info text on stringency index}
\usage{
stringency_info(sep = "<br/>")
}
\arguments{
\item{sep}{separator, break line.}
}
\description{
info text on stringency index
}
