% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_continent.R
\name{subcont_palette_calc}
\alias{subcont_palette_calc}
\title{Derives palette for continent maps and graphs based on area_map_spec}
\usage{
subcont_palette_calc(col_cont = area_map_spec(cont, "col"), x)
}
\arguments{
\item{col_cont}{named character vector area_map_spec(cont, 'col')}

\item{x}{named character vector of areas}
}
\value{
named character vector of colors for the areas
}
\description{
Derives palette for continent maps and graphs based on area_map_spec
}
