% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_SWE.R
\name{SWE}
\alias{SWE}
\title{Sweden}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/iso_SWE.R}
}
\usage{
SWE(level)
}
\description{
Sweden
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_arcgis.se.R}{Public Health Agency of Sweden}:
confirmed cases,
deaths.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_arcgis.se.R}{Public Health Agency of Sweden}:
confirmed cases.
}
}
}

\concept{Level 1}
\concept{Level 2}
