% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switzerland.R
\name{switzerland}
\alias{switzerland}
\title{Switzerland}
\usage{
switzerland(type = "state", raw = FALSE)
}
\arguments{
\item{type}{one of \code{country} (data by country) or \code{state} (data by canton). Default \code{state}, data by canton.}

\item{raw}{logical. Skip data cleaning? Default \code{FALSE}.}
}
\value{
Grouped \code{tibble} (\code{data.frame}). \href{https://github.com/emanuele-guidotti/COVID19}{More info}
}
\description{
Tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.
Swiss data by country or state (cantons).
}
\details{
See \href{https://github.com/emanuele-guidotti/COVID19}{data sources}
}
\examples{
# data by country
x <- switzerland("country")

# data by canton
x <- switzerland("state")

}
