% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_THA.R
\name{THA}
\alias{THA}
\title{Thailand}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_THA.R}
}
\usage{
THA(level)
}
\description{
Thailand
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_go.th.R}{Department of Disease Control, Thailand Ministry of Public Health}:
confirmed cases,
deaths.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/THA.csv}{Statoids (2010)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_go.th.R}{Department of Disease Control, Thailand Ministry of Public Health}:
confirmed cases,
deaths.
}
}
}

\concept{Level 1}
\concept{Level 2}
