% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{theme_simple}
\alias{theme_simple}
\title{Theme simple}
\usage{
theme_simple(base_size = 12, base_family = "Arial")
}
\arguments{
\item{base_size}{Base size}

\item{base_family}{Base font family}
}
\description{
Simple plot theme
}
\keyword{internal}
