% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_plots.R
\name{time_evol_area_facet_plot}
\alias{time_evol_area_facet_plot}
\title{Time evolution as area plot facet}
\usage{
time_evol_area_facet_plot(df, log)
}
\arguments{
\item{df}{data.frame with column called Date and x column to plot}

\item{log}{character string log or linear}
}
\value{
area plot by date
}
\description{
Time evolution as area plot facet
}
