% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{time_evol_area_plot}
\alias{time_evol_area_plot}
\title{Time evolution as area plot}
\usage{
time_evol_area_plot(
  df,
  stack = F,
  log = F,
  text = "",
  hosp = FALSE,
  active_hosp,
  formatdate = "\%d-\%m-\%y"
)
}
\arguments{
\item{df}{data.frame with column called Date and x column to plot}

\item{stack}{logical for producing a stacked plot}

\item{log}{logical for applying log scale}

\item{text}{element for tooltip}

\item{hosp}{logical, if TRUE hosp variables are in status. Default FALSE}

\item{active_hosp}{logical, if TRUE hosp and active are in status, active to be adjusted. Default FALSE}

\item{formatdate}{character a valid date format, used to remove years, default = '%d-%m-%y'}
}
\value{
area plot of given variable by date
}
\description{
Time evolution as area plot
}
\examples{
\dontrun{
df <- get_timeseries_full_data() \%>\%
get_timeseries_by_contagion_day_data \%>\%
get_timeseries_global_data()

df <- df \%>\%
pivot_longer(cols = -date, names_to = "status", values_to = "value") \%>\%
 mutate(status = as.factor(status)) \%>\%
 capitalize_names_df()

df \%>\% time_evol_line_plot()
}

\dontrun{
country_tot <- get_timeseries_full_data() \%>\%
 get_timeseries_by_contagion_day_data \%>\%
 aggregate_country_data() \%>\%
 top_n(10)

df <- get_timeseries_full_data() \%>\%
 get_timeseries_by_contagion_day_data() \%>\%
 aggregate_province_timeseries_data() \%>\%
 filter(Country.Region \%in\% country_tot$Country.Region) \%>\%
 select(Country.Region, date, confirmed)

 df <- df \%>\%
 mutate(status = as.factor(Country.Region)) \%>\%
 mutate(value = confirmed) \%>\%
 capitalize_names_df()

df \%>\% time_evol_line_plot()
}

}
