% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{time_evol_line_facet_plot}
\alias{time_evol_line_facet_plot}
\title{Time evolution as area plot facet}
\usage{
time_evol_line_facet_plot(
  df,
  log,
  g_palette = graph_palette,
  formatdate = "\%d-\%m-\%y"
)
}
\arguments{
\item{df}{data.frame with column called Date and x column to plot}

\item{log}{character string log or linear}

\item{g_palette}{character vector of colors for the graph and legend}
}
\value{
area plot by date
}
\description{
Time evolution as area plot facet
}
