% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_plots.R
\name{time_evol_line_plot}
\alias{time_evol_line_plot}
\title{Time evolution as line plot}
\usage{
time_evol_line_plot(df)
}
\arguments{
\item{df}{data.frame with column called Date and x column to plot}
}
\value{
line plot of given variable by date
}
\description{
Time evolution as line plot
}
\examples{
\dontrun{
df <- get_timeseries_full_data() \%>\%
get_timeseries_by_contagion_day_data \%>\%
get_timeseries_global_data()

df <- df \%>\%
pivot_longer(cols = -date, names_to = "status", values_to = "value") \%>\%
 mutate(status = as.factor(status)) \%>\%
 capitalize_names_df()

df \%>\% time_evol_line_plot()
}

\dontrun{
country_tot <- get_timeseries_full_data() \%>\%
 get_timeseries_by_contagion_day_data \%>\%
 aggregate_country_data() \%>\%
 top_n(10)

df <- get_timeseries_full_data() \%>\%
 get_timeseries_by_contagion_day_data() \%>\%
 aggregate_province_timeseries_data() \%>\%
 filter(Country.Region \%in\% country_tot$Country.Region) \%>\%
 select(Country.Region, date, confirmed)

 df <- df \%>\%
 mutate(status = as.factor(Country.Region)) \%>\%
 mutate(value = confirmed) \%>\%
 capitalize_names_df()

df \%>\% time_evol_line_plot()
}

}
