% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_group_plots_server.R
\name{timeline_info}
\alias{timeline_info}
\title{info text on timeline plots}
\usage{
timeline_info(sep = "<br/>", hosp = TRUE, country = "the world")
}
\arguments{
\item{sep}{separator, break line.}

\item{hosp}{logical, UCU message, default = TRUE.}

\item{country}{character, country text.}
}
\description{
info text on timeline plots
}
