% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesPlot.R
\name{timeSeriesPlot}
\alias{timeSeriesPlot}
\title{timeSeriesPlot}
\usage{
timeSeriesPlot(start_date, end_date, country_name)
}
\arguments{
\item{start_date}{start date of time range}

\item{end_date}{end date of time range}

\item{country_name}{name of specific country}
}
\description{
Time Series Plot to show distribution of death and
all infections for specific country and specific time range
}
\examples{
timeSeriesPlot("2020-10-01", "2020-10-25", "Iran")
}
