% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_toyokeizai.net.R
\name{toyokeizai.net}
\alias{toyokeizai.net}
\title{Toyo Keizai}
\source{
https://toyokeizai.net/sp/visual/tko/covid19/
}
\usage{
toyokeizai.net(level)
}
\arguments{
\item{level}{1, 2}
}
\description{
Data source for: Japan
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
}
}

\section{Level 2}{

\itemize{
\item confirmed cases
\item deaths
\item recovered
\item tests
\item hospitalizations
\item intensive care
}
}

\keyword{internal}
