% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_TWN.R
\name{TWN}
\alias{TWN}
\title{Taiwan}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_TWN.R}
}
\usage{
TWN(level)
}
\description{
Taiwan
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gov.tw.R}{Ministry of Health and Welfare of Taiwan}:
confirmed cases,
tests.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gov.tw.R}{Ministry of Health and Welfare of Taiwan}:
confirmed cases.
}
}
}

\concept{Level 1}
\concept{Level 2}
