% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.R
\name{us}
\alias{us}
\title{United States}
\usage{
us(type = "state", raw = FALSE)
}
\arguments{
\item{type}{one of \code{country} (data by country) or \code{state} (data by state). Default \code{state}, data by state.}

\item{raw}{logical. Skip data cleaning? Default \code{FALSE}.}
}
\value{
Grouped \code{tibble} (\code{data.frame}). \href{https://github.com/emanuele-guidotti/COVID19}{More info}
}
\description{
Tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.
US data by country or state.
}
\details{
See \href{https://github.com/emanuele-guidotti/COVID19}{data sources}
}
\examples{
\dontrun{
# data by country
x <- us("country")

# data by state
x <- us("state")
}
}
