% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_group_plots_server.R
\name{vaccines_info}
\alias{vaccines_info}
\title{info text on vaccine doses}
\usage{
vaccines_info(sep = "<br/>")
}
\arguments{
\item{sep}{separator, break line.}
}
\description{
info text on vaccine doses
}
