% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatterplot.R
\name{vars_growth}
\alias{vars_growth}
\title{selectInput UI parameters for growth}
\usage{
vars_growth(ax = "y")
}
\arguments{
\item{ax}{axis, x or y}
}
\description{
selectInput UI parameters for growth
}
