% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatterplot.R
\name{vars_hosp}
\alias{vars_hosp}
\title{selectInput UI parameters for hospitalization}
\usage{
vars_hosp(ax = "y")
}
\arguments{
\item{ax}{axis, x or y}
}
\description{
selectInput UI parameters for hospitalization
}
