% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{vars_nthcases_plot}
\alias{vars_nthcases_plot}
\title{Variables used in nthcase_plot}
\format{
An object of class \code{character} of length 20.
}
\usage{
vars_nthcases_plot
}
\description{
Variables used in nthcase_plot
}
\keyword{datasets}
