% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatterplot.R
\name{vars_vs_stringency_vax}
\alias{vars_vs_stringency_vax}
\title{selectInput UI parameters for stringency}
\usage{
vars_vs_stringency_vax(ax = "y")
}
\arguments{
\item{ax}{axis, x or y}
}
\description{
selectInput UI parameters for stringency
}
