% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_calc_continent.R
\name{varsNames}
\alias{varsNames}
\title{List of variable names to be used for map}
\usage{
varsNames(vars)
}
\arguments{
\item{vars}{variable name to be selected, if empty tehn all are returned}
}
\value{
list All variables, if vars is missing, or one variable.
}
\description{
List of variable names to be used for map
}
\details{
The name of the list component correspond to the variable label
}
