% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_VIR.R
\name{VIR}
\alias{VIR}
\title{Virgin Islands, U.S.}
\source{
https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/iso_VIR.R
}
\usage{
VIR(level, ...)
}
\description{
Data available at level 1 (nation).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.nytimes.covid19data.R}{The New York Times}
(confirmed cases, deaths);
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_ourworldindata.org.R}{Our World in Data}
(tests, hospitalizations, vaccines);
\href{https://data.worldbank.org/indicator/SP.POP.TOTL}{World Bank Open Data}
(population 2018).
}

\concept{level 1}
