% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_who.int.R
\name{who.int}
\alias{who.int}
\title{World Health Organization}
\source{
https://covid19.who.int
}
\usage{
who.int(level = 1, id = NULL)
}
\arguments{
\item{level}{1}

\item{id}{filter by 2-letter ISO code of country}
}
\description{
Data source for: Worldwide
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item deaths
}
}

\keyword{internal}
