% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_wikipedia.dp.R
\name{wikipedia.dp}
\alias{wikipedia.dp}
\title{Wikipedia}
\source{
https://en.wikipedia.org/wiki/COVID-19_pandemic_on_Diamond_Princess
}
\usage{
wikipedia.dp(level)
}
\arguments{
\item{level}{1}
}
\description{
Data source for: Diamond Princess (Cruise Ship)
}
\section{Level 1}{

\itemize{
\item confirmed cases
\item tests
}
}

\keyword{internal}
