% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_WLF.R
\name{WLF}
\alias{WLF}
\title{Wallis and Futuna}
\source{
https://github.com/covid19datahub/COVID19/blob/master/R/iso_WLF.R
}
\usage{
WLF(level, ...)
}
\description{
Data available at level 1 (nation).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_jhucsse_git.R}{Johns Hopkins Center for Systems Science and Engineering}
(confirmed cases, recovered, deaths);
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata_org.R}{Our World in Data}
(tests, hospitalizations, vaccines);
\href{https://data.worldbank.org/indicator/SP.POP.TOTL}{World Bank Open Data}
(population 2018).
}

\concept{level 1}
