% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write.csv}
\alias{write.csv}
\title{Data Output}
\usage{
write.csv(x, file, row.names = FALSE, na = "", fileEncoding = "UTF-8", ...)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce x to a data frame.}

\item{file}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}

\item{row.names}{either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}

\item{na}{the string to use for missing values in the data.}

\item{fileEncoding}{character string: if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded as they are written.}

\item{...}{arguments passed to \code{\link[utils:write.table]{write.csv}}}
}
\value{
return value of \code{\link[utils:write.table]{write.csv}}
}
\description{
Write csv in UTF-8.
}
