% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{y_vs_x_calc}
\alias{y_vs_x_calc}
\title{Calculates growth vs prevalence, growth vs stringency factors}
\usage{
y_vs_x_calc(data, yvar, xvar, yLab = "Growth", xLab = "Prevalence")
}
\arguments{
\item{data}{data.frame aggregated data per region}

\item{yvar}{y variable, growth factors}

\item{xvar}{x variable, prevalence over 1 M or stringency index}
}
\value{
factor vector
}
\description{
Calculates growth vs prevalence, growth vs stringency factors
}
\note{
factors created:
c('Low Growth and Prevalence', 'Low Growth - High Prevalence', 'High Growth - Low Prevalence', 'High Growth and Prevalence')
}
