% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_ZAF.R
\name{ZAF}
\alias{ZAF}
\title{South Africa}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/iso_ZAF.R}
}
\usage{
ZAF(level)
}
\description{
South Africa
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.dsfsi.covid19za.R}{Data Science for Social Impact research group, University of Pretoria}:
confirmed cases,
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.dsfsi.covid19za.R}{Data Science for Social Impact research group, University of Pretoria}:
confirmed cases,
deaths,
recovered,
tests.
}
}
}

\concept{Level 1}
\concept{Level 2}
