% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_ABW.R
\name{ABW}
\alias{ABW}
\title{Aruba}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/6482ecb1ca3fb0611afe24428190df415f41446e/R/iso_ABW.R}
}
\usage{
ABW(level)
}
\description{
Aruba
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/6482ecb1ca3fb0611afe24428190df415f41446e/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/6482ecb1ca3fb0611afe24428190df415f41446e/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
confirmed cases,
deaths,
recovered.
\item \href{https://github.com/covid19datahub/COVID19/blob/6482ecb1ca3fb0611afe24428190df415f41446e/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
