% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{add_growth_death_rate}
\alias{add_growth_death_rate}
\title{Add growth/lethality rates}
\usage{
add_growth_death_rate(df, group = "Country.Region", time = "date")
}
\arguments{
\item{df}{data.frame}

\item{group}{character Country.Region or continent or subcontinent}

\item{time}{character date}
}
\value{
df dataframe
}
\description{
Add growth/lethality rates
}
