% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_iso}
\alias{add_iso}
\title{Add XXX.csv file}
\usage{
add_iso(x, iso, ds, level, map = c("id"), append = TRUE)
}
\arguments{
\item{x}{\code{data.frame} generated by a data source function.}

\item{iso}{ISO code (3 letters).}

\item{ds}{name of the data source function generating \code{x}.}

\item{map}{named vector mapping the columns of \code{x} to the columns of the XXX.csv file.}

\item{append}{logical. Append the data to the XXX.csv file if it already exists? Default \code{TRUE}.}
}
\value{
\code{data.frame}
}
\description{
Add new country in the inst/extdata/db/ folder
}
\examples{
\dontrun{

# download data
x <- COVID19:::jhucsse_git(file = "US", cache = TRUE, level = 3, country = "USA")

# add iso
csv <- add_iso(x, iso = "USA", ds = "jhucsse_git", level = 3, map = c(
 "id"    = "id", 
 "state" = "administrative_area_level_2", 
 "city"  = "administrative_area_level_3",
 "pop"   = "population",
 "lat"   = "latitude",
 "lng"   = "longitude",
 "fips"  = "key_numeric"))
 
}

}
