% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_BRA.R
\name{BRA}
\alias{BRA}
\title{Brazil}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_BRA.R}
}
\usage{
BRA(level, ...)
}
\description{
Brazil
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.wcota.covid19br.R}{Wesley Cota}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/BRA.csv}{Instituto Brasileiro de Geografia e Estatistica - IBGE (2020)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.wcota.covid19br.R}{Wesley Cota}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/BRA.csv}{Wesley Cota (2020)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.wcota.covid19br.R}{Wesley Cota}:
confirmed cases,
deaths.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.eguidotti.covid19br.R}{Emanuele Guidotti}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
