% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_CAC.R
\name{CAC}
\alias{CAC}
\title{Cruise Ship Costa Atlantica}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_CAC.R}
}
\usage{
CAC(level)
}
\description{
Cruise Ship Costa Atlantica
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.swsoyee.2019ncovjapan.R}{Su Wei}:
confirmed cases,
deaths,
recovered,
tests,
hospitalizations.
}
}
}

\concept{Level 1}
