% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_ceeds_git.R
\name{ceeds_git}
\alias{ceeds_git}
\title{Centre of Excellence in Economics and Data Science, University of Milan}
\source{
https://github.com/CEEDS-DEMM/COVID-Pro-Dataset
}
\usage{
ceeds_git(level)
}
\description{
Imports Italian deaths at the province level from the Centre of Excellence
in Economics and Data Science, University of Milan.
}
\keyword{internal}
