% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CH}
\alias{CH}
\title{Demographic indicators - Switzerland}
\format{\code{data.frame}}
\source{
\url{https://www.bfs.admin.ch}
}
\usage{
CH
}
\description{
Tidy format dataset of Swiss demographic indicators from the
\href{https://www.bfs.admin.ch/bfs/en/home/statistics/regional-statistics/regional-portraits-key-figures/cantons/data-explanations.html}{Swiss Federal Statistical Office}.
}
\details{
The columns \code{pop_14}, \code{pop_15_64} and \code{pop_65} correspond to the fraction of people
under age 19, between 20 and 64, or over 65. The labels \code{pop_14}, \code{pop_15_64} and \code{pop_65} are kept
for compatibility with the datasets \code{\link{WB}}, \code{\link{IT}}.
}
\seealso{
\code{\link{WB}}, \code{\link{IT}}
}
\keyword{datasets}
