% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_CHL.R
\name{CHL}
\alias{CHL}
\title{Chile}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/iso_CHL.R}
}
\usage{
CHL(level)
}
\description{
Chile
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.minciencia.datoscovid19.R}{Ministerio de Ciencia, Tecnología, Conocimiento, e Innovación}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care,
patients requiring ventilation.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.minciencia.datoscovid19.R}{Ministerio de Ciencia, Tecnología, Conocimiento, e Innovación}:
confirmed cases,
deaths,
recovered,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
intensive care.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.minciencia.datoscovid19.R}{Ministerio de Ciencia, Tecnología, Conocimiento, e Innovación}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
