% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_CHN.R
\name{CHN}
\alias{CHN}
\title{China}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_CHN.R}
}
\usage{
CHN(level)
}
\description{
China
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_who.int.R}{World Health Organization}:
confirmed cases,
deaths.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
