% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_calc_continent.R
\name{choose_domain}
\alias{choose_domain}
\title{Utility to choose domain for legend}
\usage{
choose_domain(x, var)
}
\arguments{
\item{x}{numeric vector of map data}

\item{var}{character vector of variable name}
}
\value{
numeric vector of range
}
\description{
Utility to choose domain for legend
}
