% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-covid19.R
\name{Covid}
\alias{Covid}
\title{R6 Covid-19 Class}
\description{
This object type is used to retrieve, analyse and plot international data on
COVID19 outbreak and spreading.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Covid$new()}}
\item \href{#method-print}{\code{Covid$print()}}
\item \href{#method-get}{\code{Covid$get()}}
\item \href{#method-regions}{\code{Covid$regions()}}
\item \href{#method-dates}{\code{Covid$dates()}}
\item \href{#method-update}{\code{Covid$update()}}
\item \href{#method-sir}{\code{Covid$sir()}}
\item \href{#method-table}{\code{Covid$table()}}
\item \href{#method-plot_hist}{\code{Covid$plot_hist()}}
\item \href{#method-plot_sir}{\code{Covid$plot_sir()}}
\item \href{#method-clone}{\code{Covid$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new Covid object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
This is to print the object type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\subsection{Method \code{get()}}{
This returns the table with raw data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$get(region = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{is a string; the name of the region to be returned.
Use "ITA" to return the national aggregated data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regions"></a>}}
\subsection{Method \code{regions()}}{
The list of all available regions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$regions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dates"></a>}}
\subsection{Method \code{dates()}}{
The list of dates available to fit a regression model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$dates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\subsection{Method \code{update()}}{
This updates the inner table downloading data from the remote repo.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sir"></a>}}
\subsection{Method \code{sir()}}{
This function implements a forecast based on a time-dependent SIR
model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$sir(
  region = "ITA",
  fit_date = max(self$dates()),
  end_date = as.Date("2020-05-31")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{is a string; the name of the region to be returned.
Use "ITA" to return the national aggregated data.}

\item{\code{fit_date}}{is a date; it is the end of the historical time series used
to produce the forecasts.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table"></a>}}
\subsection{Method \code{table()}}{
This is used to export a plottable table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$table(region = "ITA")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{is a string; the name of the region to be returned.
Use "ITA" to return the national aggregated data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_hist"></a>}}
\subsection{Method \code{plot_hist()}}{
This plots a barchart of the summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$plot_hist(region = "ITA", log = FALSE, delta = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{is a string; the name of the region to be returned.
Use "ITA" to return the national aggregated data.}

\item{\code{log}}{is a boolean; if \code{TRUE}, a logarithmic scale is applied to y
axis in plots.}

\item{\code{delta}}{is a boolean; if \code{TRUE}, the difference from the
previous day is shown.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_sir"></a>}}
\subsection{Method \code{plot_sir()}}{
This plots a linechart of the forecast.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$plot_sir(
  region = "ITA",
  fit_date = max(self$dates()),
  end_date = as.Date("2020-05-31")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{region}}{is a string; the name of the region to be returned.
Use "ITA" to return the national aggregated data.}

\item{\code{fit_date}}{is a date; it is the end of the historical time series used
to produce the forecasts.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Covid$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
