% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{covid19}
\alias{covid19}
\title{Process and Clean COVID-19 Raw Data}
\usage{
covid19(x, raw = FALSE)
}
\arguments{
\item{x}{COVID-19 \code{data.frame}}

\item{raw}{logical. Skip data cleaning? Default \code{FALSE}.}
}
\value{
Tidy format \code{tibble} (\code{data.frame}) grouped by id, containing the columns:
\describe{
 \item{id}{id in the form "country|state|city".}
 \item{date}{date.}
 \item{country}{administrative area level 1.}
 \item{state}{administrative area level 2.}
 \item{city}{administrative area level 3.}
 \item{lat}{latitude.}
 \item{lng}{longitude.}
 \item{deaths}{the number of deaths.}
 \item{confirmed}{the number of cases.}
 \item{tests}{the number of tests.}
 \item{deaths_new}{daily increase in the number of deaths.}
 \item{confirmed_new}{daily increase in the number of cases.}
 \item{tests_new}{daily increase in the number of tests.}
 \item{pop}{total population.}
 \item{pop_14}{population ages 0-14 (\% of total population). Except Switzerland: ages 0-19.}
 \item{pop_15_64}{population ages 15-64 (\% of total population). Except Switzerland: ages 20-64.}
 \item{pop_65}{population ages 65+ (\% of total population).}
 \item{pop_age}{median age of population.}
 \item{pop_density}{population density per km2.}
 \item{pop_death_rate}{population mortality rate.}
}
}
\description{
Internal function used to clean the raw data. Provides a
unified and curated COVID-19 dataset across different sources.
}
