% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19.R
\name{COVID19}
\alias{COVID19}
\title{Package Update}
\source{
\href{https://github.com/emanuele-guidotti/COVID19}{GitHub repository}
}
\usage{
COVID19()
}
\value{
install and require the latest development version of \code{COVID19}.
}
\description{
Updates the package to the latest development version.
}
