% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{covid19}
\alias{covid19}
\title{COVID-19 Data Hub}
\source{
\url{https://covid19datahub.io}
}
\usage{
covid19(country = NULL, level = 1)
}
\arguments{
\item{country}{vector of country names or \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{ISO codes} (alpha-2, alpha-3 or numeric).}

\item{level}{integer. Granularity level. 1: country-level data. 2: state-level data. 3: city-level data.}
}
\description{
COVID-19 Data Hub
}
\references{
Guidotti, E., Ardia, D., (2020), "COVID-19 Data Hub", Journal of Open Source Software 5(51):2376, \doi{10.21105/joss.02376}.
}
\keyword{internal}
