% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{covid19}
\alias{covid19}
\title{Coronavirus COVID-19 (2019-nCoV) Epidemic Datasets}
\source{
\href{https://github.com/covid19datahub/COVID19#data-sources}{Data sources}
}
\usage{
covid19(
  ISO = NULL,
  level = 1,
  start = "2019-01-01",
  end = Sys.Date(),
  vintage = FALSE,
  raw = FALSE,
  cache = TRUE
)
}
\arguments{
\item{ISO}{vector of ISO codes to retrieve (alpha-2, alpha-3 or numeric). Each country is identified by one of its \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{ISO codes}}

\item{level}{integer. Granularity level. 1: country-level data. 2: state-level data. 3: city-level data.}

\item{start}{the start date of the period of interest.}

\item{end}{the end date of the period of interest.}

\item{vintage}{logical. Retrieve the snapshot of the dataset at the \code{end} date instead of using the latest version? Default \code{FALSE}.}

\item{raw}{logical. Skip data cleaning? Default \code{FALSE}. See details.}

\item{cache}{logical. Memory caching? Significantly improves performance on successive calls. Default \code{TRUE}.}
}
\value{
\href{https://github.com/covid19datahub/COVID19#dataset}{Grouped \code{tibble} (\code{data.frame})}
}
\description{
Unified tidy format datasets of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic across several sources.
The data are downloaded in real-time, cleaned and matched with exogenous variables.
}
\details{
\href{https://github.com/covid19datahub/COVID19}{Collection methodology and details}
}
\examples{
\dontrun{

# Worldwide data by country
covid19()

# Worldwide data by state
covid19(level = 2)

# US data by state
covid19("USA", level = 2)

# Swiss data by state (cantons)
covid19("CHE", level = 2)

# Italian data by state (regions)
covid19("ITA", level = 2)

# Italian and US data by city
covid19(c("ITA","USA"), level = 3)
}

}
