% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.m3it.covid19data.R
\name{github.m3it.covid19data}
\alias{github.m3it.covid19data}
\title{Matt Bolton}
\source{
https://github.com/M3IT/COVID-19_Data
}
\usage{
github.m3it.covid19data(level)
}
\description{
Imports confirmed cases, deaths, recovered, tests, hospitalizations, and vaccines
from Matt Bolton's Github repository for Australia. Data aggregated from www.covid19data.com.au,
which is sourced from Australian governments at federal, state and territory levels.
See www.covid19data.com.au/data-notes for details and notes.
}
\keyword{internal}
