% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_covid19vaccini_git.R
\name{covid19vaccini_git}
\alias{covid19vaccini_git}
\title{Commissario straordinario per l'emergenza Covid-19, Presidenza del Consiglio dei Ministri}
\source{
github.com/italia/covid19-opendata-vaccini
}
\usage{
covid19vaccini_git(level)
}
\description{
Imports Italian vaccination data at national and regional level from Commissario
straordinario per l'emergenza Covid-19.
}
\details{
Italy provides vaccination data as follows:
\itemize{
\item number of first and second doses for a 2 doses vaccination cycle
\item number of oneshot doses for a 1 dose vaccination cycle
\item number of extra doses
}

We compute vaccines_1 as first+oneshot and vaccines_2 as second+oneshot..
}
\keyword{internal}
