% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cumsum}
\alias{cumsum}
\title{Cumulative Sums}
\usage{
cumsum(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric object.}

\item{na.rm}{logical. Should missing values be removed? Default \code{FALSE}.}
}
\description{
Returns a numeric object whose elements are the cumulative sums of the elements of the argument.
}
\details{
If \code{na.rm=TRUE}, then \code{NA} are treated as \code{0} when computing the cumulative sum.
}
\examples{
\dontrun{

x <- mtcars[1:5,]
x[2,] <- NA

cumsum(x)
cumsum(x, na.rm = TRUE)

}

}
