% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{docstring}
\alias{docstring}
\title{Generate docstring to use in the iso_ files to list the population data source}
\usage{
docstring(iso, level)
}
\arguments{
\item{iso}{the ISO code of the country}

\item{level}{1, 2, 3}
}
\description{
Generate docstring to use in the iso_ files to list the population data source
}
\keyword{internal}
