% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_check_format}
\alias{ds_check_format}
\title{Check Data Source Format}
\usage{
ds_check_format(x, level)
}
\arguments{
\item{x}{output of a data source function.}

\item{level}{integer. Granularity level. 1: country-level data. 2: state-level data. 3: city-level data.}
}
\value{
logical.
}
\description{
Checks if the output of a data source function is correctly formatted. 
The function checks the FORMAT, NOT the DATA.
Before submission, the data should be double checked by comparing with external data sources (e.g. Google search).
}
\examples{
\dontrun{

# check format of the 'pcmdpc_git' data source
x <- COVID19:::pcmdpc_git(cache = FALSE, level = 1)
ds_check_format(x, level = 1)

}

}
