% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extdata}
\alias{extdata}
\title{External Data}
\usage{
extdata(...)
}
\arguments{
\item{...}{path to file}
}
\value{
\code{data.frame}
}
\description{
Read files in the inst/extdata/ folder.
}
\examples{
\dontrun{

# read file inst/extdata/db/ISO.csv
x <- extdata("db","ISO.csv")

}

}
