% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_FRA.R
\name{FRA}
\alias{FRA}
\title{France}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/iso_FRA.R}
}
\usage{
FRA(level)
}
\description{
France
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/ds_gouv.fr.R}{Santé Publique France}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}:
recovered.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/ds_gouv.fr.R}{Santé Publique France}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}


\subsection{Level 3}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/ds_gouv.fr.R}{Santé Publique France}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
\concept{Level 2}
\concept{Level 3}
