% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{from_contagion_day_bar_facet_plot}
\alias{from_contagion_day_bar_facet_plot}
\title{Evolution from contagion day facet}
\usage{
from_contagion_day_bar_facet_plot(
  df,
  xdate = "date",
  formatdate = "\%d-\%m-\%y"
)
}
\arguments{
\item{df}{data.frame to plot}

\item{xdate}{character variable for x axis}
}
\value{
barplot facet
}
\description{
Evolution from contagion day facet
}
