% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_datahub}
\alias{get_datahub}
\title{Get timeseries full data from datahub}
\usage{
get_datahub(stardate = "2020-01-15", lev = 1, verbose = FALSE)
}
\arguments{
\item{stardate}{character staring date}

\item{lev}{integer 1 for country level, 2 for reagions}
}
\value{
data tibble of confirmed, deaths, active and recovered Country.Region
}
\description{
Get timeseries full data from datahub
}
\details{
data sourced from https://github.com/covid19datahub/COVID19/
}
