% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data.R, R/get_data.R
\name{build_data}
\alias{build_data}
\alias{get_datahub_fix_ch}
\alias{get_datahub}
\title{Build data in GutHub action yml and save as RDS}
\usage{
build_data()

get_datahub_fix_ch(
  country = NULL,
  startdate = "2020-01-22",
  lev = 1,
  verbose = FALSE
)

get_datahub(
  country = NULL,
  startdate = "2020-01-22",
  lev = 1,
  verbose = FALSE,
  hosp = TRUE,
  cache = FALSE
)
}
\arguments{
\item{country}{character country, to chose with lev = 2}

\item{startdate}{character staring date}

\item{lev}{integer 1 for country level, 2 for regions}

\item{verbose}{logical. Print data sources? Default FALSE (opposite from \code{covid19})}

\item{hosp}{logical. If TRUE hospitalized detailed data are retrieved. Default TRUE since release 2.3.1}

\item{cache}{logical. If TRUE cache argument is used in the covid19() call}
}
\value{
data tibble of confirmed, deaths, active and recovered etc Country.Region

data tibble of confirmed, deaths, active and recovered Country.Region
}
\description{
Build data in GutHub action yml and save as RDS

Get timeseries full data from datahub adding CH hospitalised data from level 2

Get timeseries full data from datahub
}
\details{
data sourced from https://github.com/covid19datahub/COVID19/

data sourced from https://github.com/covid19datahub/COVID19/
}
