% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_GLP.R
\name{GLP}
\alias{GLP}
\title{Guadeloupe}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_GLP.R}
}
\usage{
GLP(level)
}
\description{
Guadeloupe
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gouv.fr.R}{Santé Publique France}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
