% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_GLP.R
\name{GLP}
\alias{GLP}
\title{Guadeloupe}
\source{
https://github.com/covid19datahub/COVID19/blob/6482ecb1ca3fb0611afe24428190df415f41446e/R/iso_GLP.R
}
\usage{
GLP(level, ...)
}
\description{
Data available at level 1 (nation).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/6482ecb1ca3fb0611afe24428190df415f41446e/R/ds_gouv_fr.R}{Santé publique France}
(confirmed cases, deaths, tests, hospitalizations, vaccines);
\href{https://data.worldbank.org/indicator/SP.POP.TOTL}{World Bank Open Data}
(population 2018).
}

\concept{level 1}
