% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_GUF.R
\name{GUF}
\alias{GUF}
\title{French Guiana}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/iso_GUF.R}
}
\usage{
GUF(level)
}
\description{
French Guiana
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_gouv.fr.R}{Santé Publique France}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
