% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_gv.at.R
\name{gv.at}
\alias{gv.at}
\title{Federal Ministry of Social Affairs, Health, Care and Consumer Protection, Austria (BMSGPK)}
\source{
https://www.data.gv.at/covid-19/
}
\usage{
gv.at(level)
}
\description{
Imports confirmed cases, deaths, recovered, tests, hospitalizations, and vaccines at
national and state level for Austria from BMSGPK. Confirmed cases, recovered, and deaths are
also available at the district level.
}
\details{
We compute vaccines_1 as the number of first doses administrated and
vaccines_2 as the number of second doses administrated (fully vaccinated).
}
\keyword{internal}
