% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_healthdata_gov.R
\name{healthdata_gov}
\alias{healthdata_gov}
\title{U.S. Department of Health & Human Services}
\source{
https://healthdata.gov/browse?tags=hhs+covid-19
}
\usage{
healthdata_gov(level)
}
\description{
Imports hospitalization data and tests for United States at level 2 from
the U.S. Department of Health & Human Services.
}
\keyword{internal}
