% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IT}
\alias{IT}
\title{Datasets}
\format{An object of class \code{data.frame} with 129 rows and 8 columns.}
\source{
https://www.istat.it/en/population-and-households?data-and-indicators
}
\usage{
IT
}
\description{
Datasets
}
\keyword{datasets}
