% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy.R
\name{italy}
\alias{italy}
\title{Coronavirus COVID-19 data - Italy}
\usage{
italy(type = "state")
}
\arguments{
\item{type}{one of \code{country} (data by country), \code{state} (data by region) or \code{city} (data by city). Default \code{state}, data by region.}
}
\value{
Return of the internal function \code{\link{covid19}}
}
\description{
Tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.
Italian data by country, state (regions) or city.
The data are downloaded in real-time, processed and merged with demographic indicators (\code{\link{IT}}).
}
\details{
Data pulled from the \href{https://github.com/pcm-dpc/COVID-19}{repository} for the
2019 Novel Coronavirus by Ministero della Salute, \href{http://www.protezionecivile.it/}{Dipartimento della Protezione Civile, Italia}.
}
\examples{
# data by country
x <- italy("country")

# data by region
x <- italy("state")

# data by city
x <- italy("city")

}
\seealso{
\code{\link{world}}, \code{\link{diamond}}, \code{\link{switzerland}}, \code{\link{liechtenstein}}
}
