% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/italy.R
\name{italy}
\alias{italy}
\title{Coronavirus COVID-19 Data - Italy}
\source{
\url{https://github.com/pcm-dpc/COVID-19}
}
\usage{
italy(type = "country")
}
\arguments{
\item{type}{one of \code{"country"} (data by country), \code{"region"} (data by region) or \code{"city"} (data by city). Default \code{"country"}.}
}
\value{
data.frame. See \url{https://github.com/pcm-dpc/COVID-19}
}
\description{
Download the Italian COVID-19 dataset from the official repository of Dipartimento della Protezione Civile.
}
\details{
See \url{https://github.com/pcm-dpc/COVID-19}
}
\examples{
\dontrun{

# data by country
x <- italy()

# data by region
x <- italy("region")

# data by city
x <- italy("city")

}

}
