% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.cssegisanddata.covid19.R
\name{github.cssegisanddata.covid19}
\alias{github.cssegisanddata.covid19}
\title{Johns Hopkins Center for Systems Science and Engineering}
\source{
https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data/csse_covid_19_time_series
}
\usage{
github.cssegisanddata.covid19(file, cache = TRUE, level = 1, country = NULL, state = NULL)
}
\description{
Imports worldwide confirmed cases, recovered, and deaths at level 1 and 2 from JHU CSSE.
Confirmed cases and deaths are also available at level 3 for United States.
}
\keyword{internal}
