% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.cssegisanddata.covid19unified.R
\name{github.cssegisanddata.covid19unified}
\alias{github.cssegisanddata.covid19unified}
\title{Johns Hopkins Center for Systems Science and Engineering}
\source{
https://github.com/CSSEGISandData/COVID-19_Unified-Dataset
}
\usage{
github.cssegisanddata.covid19unified(iso, level)
}
\description{
Imports worldwide confirmed cases, recovered, deaths, tests, and hospitalizations at various
levels from the JHU CSSE Unified Dataset.
}
\keyword{internal}
