% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_JPN.R
\name{JPN}
\alias{JPN}
\title{Japan}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/iso_JPN.R}
}
\usage{
JPN(level)
}
\description{
Japan
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.swsoyee.2019ncovjapan.R}{Su Wei}:
confirmed cases,
deaths,
recovered,
tests,
hospitalizations,
intensive care,
patients requiring ventilation.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_ourworldindata.org.R}{Our World in Data}:
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated.
}
}


\subsection{Level 2}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.swsoyee.2019ncovjapan.R}{Su Wei}:
confirmed cases,
deaths,
recovered,
tests,
hospitalizations,
intensive care,
patients requiring ventilation.
}
}
}

\concept{Level 1}
\concept{Level 2}
