% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_countries_data}
\alias{load_countries_data}
\title{load countries  data}
\usage{
load_countries_data(destpath = "./inst/countries_data/")
}
\arguments{
\item{destpath}{path to file}
}
\value{
countries shapefile
}
\description{
load countries  data
}
