% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lw_vars_calc}
\alias{lw_vars_calc}
\title{Computes last week variables from \code{build_data_aggr}}
\usage{
lw_vars_calc(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame withe last week variables added
}
\description{
Computes last week variables from \code{build_data_aggr}
}
\note{
Last week variables have prefix 'lw'
}
