% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_values}
\alias{map_values}
\title{Map values}
\usage{
map_values(x, map)
}
\arguments{
\item{x}{vector.}

\item{map}{named vector mapping names to values.}
}
\value{
\code{vector}.
}
\description{
Map values of a vector.
}
\examples{
\dontrun{

x <- c('red','green','red','blue')

map_values(x, map = c(
'red' = 'yellow', 
'blue' = 'orange'))

}

}
