% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_MNP.R
\name{MNP}
\alias{MNP}
\title{Northern Mariana Islands}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_MNP.R}
}
\usage{
MNP(level)
}
\description{
Northern Mariana Islands
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.nytimes.covid19data.R}{The New York Times}:
confirmed cases,
deaths.
}
}
}

\concept{Level 1}
