% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_caseBoxes.R
\name{mod_caseBoxes}
\alias{mod_caseBoxes}
\alias{mod_caseBoxes_ui}
\alias{mod_caseBoxes_server}
\title{caseBoxes UI Function}
\usage{
mod_caseBoxes_ui(id, hosp = FALSE)

mod_caseBoxes_server(input, output, session, counts, hosp = FALSE, vax = NULL)
}
\arguments{
\item{id, }{Internal parameters for {shiny}.}

\item{hosp}{logical, if TRUE hospitalised box variables is added.}

\item{counts}{Reactive expression yielding the named vector of cases by type.}
}
\description{
A shiny Module for displaying the number of cases by type as
  colored boxes.
}
\examples{
if (interactive()) {
  devtools::load_all()
  long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
  ui <- fluidPage(
    tagList(
      Covid19Mirai:::golem_add_external_resources(),
      Covid19Mirai:::mod_caseBoxes_ui("boxes")
    )
  )
  server <- function(input, output, session) {
    statuses <- c("confirmed", "deaths", "recovered", "active")
    # select all variables
    allstatuses = c(statuses, paste0("new_", statuses))
    counts <- reactive({
      invalidateLater(2000)
      as.list(stats::setNames(sample.int(1000, length(allstatuses)), allstatuses))
    })
    callModule(Covid19Mirai:::mod_caseBoxes_server, "boxes", counts = counts)
  }
  runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
}

}
\keyword{internal}
