% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_caseBoxes.R
\name{mod_caseBoxes}
\alias{mod_caseBoxes}
\alias{mod_caseBoxes_ui}
\alias{mod_caseBoxes_server}
\title{caseBoxes UI Function}
\usage{
mod_caseBoxes_ui(id, hosp = FALSE)

mod_caseBoxes_server(input, output, session, counts, hosp = FALSE, vax = NULL)
}
\arguments{
\item{id, }{Internal parameters for {shiny}.}

\item{hosp}{logical, if TRUE hospitalised box variables is added.}

\item{counts}{Reactive expression yielding the named vector of cases by type.}
}
\description{
A shiny Module for displaying the number of cases by type as
  colored boxes.
}
\examples{
if (interactive()) {
 # devtools::load_all()
  long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
  ui <- fluidPage(
    tagList(
      Covid19Mirai:::golem_add_external_resources(),
      Covid19Mirai:::mod_caseBoxes_ui("boxes")
    )
  )
  server <- function(input, output, session) {

    orig_data <-get_datahub() \%>\%
      get_timeseries_by_contagion_day_data()

    pop_data = get_pop_datahub()
    orig_data_aggregate = build_data_aggr(orig_data, pop_data)

    country_data = orig_data_aggregate \%>\%
      filter(Country.Region == "France")

    country_data_today <- country_data \%>\%
      filter(date == maxdate)

    lw_country_data = lw_vars_calc(country_data)
    pw_country_data =  lw_vars_calc(country_data, 14)

    country_data_today = country_data_today \%>\%
      left_join(lw_country_data \%>\% select(-population))  \%>\%
      left_join(pw_country_data \%>\% select(-population))

    vaxarg = "recovered"

    callModule(mod_caseBoxes_server, "boxes", counts = country_data_today, vax = vaxarg)
  }
  runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
}

if (interactive()) {
  # devtools::load_all()
  long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
  ui <- fluidPage(
    tagList(
      Covid19Mirai:::golem_add_external_resources(),
      Covid19Mirai:::mod_caseBoxes_ui("boxes", hosp = TRUE)
    )
  )
  server <- function(input, output, session) {

    orig_data <-get_datahub() \%>\%
      get_timeseries_by_contagion_day_data()

    pop_data = get_pop_datahub()
    orig_data_aggregate = build_data_aggr(orig_data, pop_data)

    country_data = orig_data_aggregate \%>\%
      filter(Country.Region == "France")

    country_data_today <- country_data \%>\%
      filter(date == maxdate)

    lw_country_data = lw_vars_calc(country_data)
    pw_country_data =  lw_vars_calc(country_data, 14)

    country_data_today = country_data_today \%>\%
      left_join(lw_country_data \%>\% select(-population))  \%>\%
      left_join(pw_country_data \%>\% select(-population))


    callModule(mod_caseBoxes_server, "boxes", counts = country_data_today, hosp = TRUE)
  }
  runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
}
}
\keyword{internal}
