% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vaccines_text.R
\name{mod_vaccines_text_ui}
\alias{mod_vaccines_text_ui}
\alias{mod_novaccines_text_server}
\alias{mod_vaccines_text_server}
\title{caseBoxes UI Function}
\usage{
mod_vaccines_text_ui(id)

mod_novaccines_text_server(input, output, session, country, what = "Vaccines")

mod_vaccines_text_server(input, output, session, df, dftoday)
}
\arguments{
\item{id, }{Internal parameters for {shiny}.}

\item{input, output, session}{Internal parameters for {shiny}.}

\item{country}{character country name}

\item{what}{variable name}

\item{df}{dataset whole time series.}

\item{dftoday}{dataset with today and weekly data.}
}
\description{
A shiny Module for displaying stats about vaccination
}
