% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_MYT.R
\name{MYT}
\alias{MYT}
\title{Mayotte}
\source{
https://github.com/covid19datahub/COVID19/blob/master/R/iso_MYT.R
}
\usage{
MYT(level, ...)
}
\description{
Data available at level 1 (nation).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_gouv_fr.R}{Santé publique France}
(confirmed cases, deaths, tests, hospitalizations, vaccines);
\href{https://data.worldbank.org/indicator/SP.POP.TOTL}{World Bank Open Data}
(population 2018).
}

