% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_MYT.R
\name{MYT}
\alias{MYT}
\title{Mayotte}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/iso_MYT.R}
}
\usage{
MYT(level)
}
\description{
Mayotte
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/ba52b2c666648c8a84f5ad3af3f5236924545608/R/ds_gouv.fr.R}{Santé Publique France}:
confirmed cases,
deaths,
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
