% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_github.nytimes.covid19data.R
\name{github.nytimes.covid19data}
\alias{github.nytimes.covid19data}
\title{The New York Times}
\source{
https://github.com/nytimes/covid-19-data
}
\usage{
github.nytimes.covid19data(level, cache = FALSE, fips = NULL)
}
\description{
Imports confirmed cases and deaths for United States at national, state, and county level
from The New York Times.
}
\keyword{internal}
