% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_ourworldindata_org.R
\name{ourworldindata_org}
\alias{ourworldindata_org}
\title{Our World in Data}
\source{
https://covid.ourworldindata.org
}
\usage{
ourworldindata_org(id = NULL)
}
\description{
Imports worldwide confirmed cases, deaths, hospitalizations, intensive care,
tests, and vaccines at national level from "Our World in Data".
}
\keyword{internal}
