% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_ourworldindata.org.R
\name{ourworldindata.org}
\alias{ourworldindata.org}
\title{Our World in Data}
\source{
https://covid.ourworldindata.org
}
\usage{
ourworldindata.org(level = 1, id = NULL)
}
\description{
Imports worldwide confirmed cases, deaths, hospitalizations,
tests, and vaccines at national level from Our World in Data.
Vaccination data are also available at state level for U.S.
}
\keyword{internal}
