% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map_calc_continent.R
\name{pal_fun_calc}
\alias{pal_fun_calc}
\title{Utility calculate colors given palette}
\usage{
pal_fun_calc(x)
}
\arguments{
\item{x}{numeric vector of map data}
}
\value{
rescaled x
}
\description{
Utility calculate colors given palette
}
