% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_day_zero.R
\name{plot_day_zero}
\alias{plot_day_zero}
\title{Plot day zero}
\usage{
plot_day_zero(
  countries = c("Italy", "Spain", "US", "Germany"),
  ylog = TRUE,
  day0 = 150,
  cumulative = TRUE,
  time_before = 0,
  add_markers = FALSE,
  line_size = 1.5,
  point_size = 1.5,
  point_alpha = 0.9,
  max_date = Sys.Date(),
  calendar = FALSE,
  deaths = FALSE,
  pop = FALSE,
  pop_adjustor = 1e+06,
  by_district = FALSE,
  districts = NULL,
  alpha = 0.8,
  roll = 0,
  roll_fun = "mean",
  color_var = "geo"
)
}
\arguments{
\item{countries}{Character vector of country names}

\item{ylog}{Whether the y-axis should be on log scale}

\item{day0}{An integer, the number of cumulative cases at which the outbreak is considered to have started}

\item{cumulative}{Whether to count cases cumulatively}

\item{time_before}{How many days before outbreak to show}

\item{line_size}{Size of line}

\item{point_size}{Size of points}

\item{point_alpha}{Alpha of points}

\item{max_date}{The maximum date}

\item{calendar}{Whether to plot by calendar date}

\item{deaths}{Whether to show deaths instead of cases}

\item{pop}{Adjust by population}

\item{pop_adjustor}{Deaths per x (default million) when doing population adjustment}

\item{by_district}{Show by district}

\item{districts}{Show by districts, default FALSE}

\item{alpha}{alpha of lines}

\item{roll}{How many days to get rolling values for}

\item{roll_fun}{Mean or Sum}

\item{add_markets}{Whether to show lines / circle at outbreak start}
}
\description{
Generate a plot with time adjusted for the day at which the outbreak is considered to have started
}
