% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rate_hist}
\alias{plot_rate_hist}
\title{plot rate as hist}
\usage{
plot_rate_hist(df, percent = F, y_min = 0, g_palette)
}
\arguments{
\item{df}{data.frame}

\item{percent}{logical to make the y axis in percent}

\item{y_min}{min value on y axis}

\item{g_palette}{character vector of colors for the graph and legend}
}
\value{
ggplot plot
}
\description{
plot rate as hist
}
