% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rate_hist}
\alias{plot_rate_hist}
\title{plot rate as hist}
\usage{
plot_rate_hist(df, color)
}
\arguments{
\item{df}{data.frame}

\item{color}{string used to define color}
}
\description{
plot rate as hist
}
