% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_PRI.R
\name{PRI}
\alias{PRI}
\title{Puerto Rico}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/master/R/iso_PRI.R}
}
\usage{
PRI(level)
}
\description{
Puerto Rico
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/master/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_github.nytimes.covid19data.R}{The New York Times}:
confirmed cases,
deaths.
\item \href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
