% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_PRI.R
\name{PRI}
\alias{PRI}
\title{Puerto Rico}
\source{
https://github.com/covid19datahub/COVID19/blob/master/R/iso_PRI.R
}
\usage{
PRI(level, ...)
}
\description{
Data available at level 1 (nation).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_nytimes_git.R}{The New York Times}
(confirmed cases, deaths);
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata_org.R}{Our World in Data}
(tests, hospitalizations, vaccines);
\href{https://data.worldbank.org/indicator/SP.POP.TOTL}{World Bank Open Data}
(population 2018).
}

