% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rate_calc}
\alias{rate_calc}
\title{Calculates rates between variables}
\usage{
rate_calc(x, y, cap = 1, floor = 0, digits = 4, nawith0 = FALSE)
}
\arguments{
\item{x}{numeric numerator}

\item{y}{numeric denominator}

\item{cap}{numeric cap, 1 or Inf for no cap}

\item{floor}{numeric floor, 0 or -Inf for no floor}

\item{digits}{numeric digits for rounding}

\item{nawith0}{if TRUE then NAs are replaced with 0s}
}
\value{
x numeric rate
}
\description{
Calculates rates between variables
}
