% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.zip}
\alias{read.zip}
\title{Data Input (zip)}
\usage{
read.zip(zip, files, cache = FALSE, ...)
}
\arguments{
\item{zip}{path (url) to the zip folder.}

\item{files}{vector of filenames to read inside the zip folder.}

\item{cache}{logical. Memory caching? Default \code{FALSE}.}

\item{...}{arguments passed to \code{\link{read.csv}} or \code{\link{read.excel}}.}
}
\value{
list of \code{data.frames}
}
\description{
Reads files from a zip folder.
}
\examples{
\dontrun{

url <- "https://info.gesundheitsministerium.at/data/data.zip"

x   <- read.zip(url, cache = TRUE, sep = ";", files = c(
"confirmed" = "Epikurve.csv",
"deaths"    = "TodesfaelleTimeline.csv",
"recovered" = "GenesenTimeline.csv"))

}

}
