% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_ecdc.R
\name{read_cases_ecdc}
\alias{read_cases_ecdc}
\title{Reads COVID-19 cases and deaths from the ECDC}
\usage{
read_cases_ecdc()
}
\value{
\code{tbl} with 11 columns:
\itemize{
\item \code{date}
\item \code{dateRep}
\item \code{day}
\item \code{month}
\item \code{year}
\item \code{cases}
\item \code{deaths}
\item \code{countriesAndTerritories}
\item \code{geoId}
\item \code{countryterritoryCode}
\item \code{popData2018}
}
}
\description{
Reads a \href{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}{dataset} with COVID-19 cases and deaths, which is maintained by the European Centre for Disease Prevention and Control (ECDC).
}
\details{
\itemize{
\item Reads an \href{https://www.ecdc.europa.eu/}{ECDC} \href{https://www.ecdc.europa.eu/en/publications-data/download-todays-data-geographic-distribution-covid-19-cases-worldwide}{dataset} of  \href{https://www.who.int/emergencies/diseases/novel-coronavirus-2019}{COVID-19} cases and deaths.
\item Adds a \code{date} column of class \code{Date}, derived from the \code{dateRep} column, which is a string.
}
}
