% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{repo}
\alias{repo}
\title{Generate permalink to the file at the GitHub repository}
\usage{
repo(x, csv = FALSE)
}
\arguments{
\item{x}{name of the iso_ or ds_ function, or name of the .csv file}
}
\description{
Generate permalink to the file at the GitHub repository
}
\keyword{internal}
