% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_up}
\alias{round_up}
\title{Rounds up numbers for labels in plots}
\usage{
round_up(maxv, down = FALSE)
}
\arguments{
\item{maxv}{numeric max value}

\item{down}{logical if TRUE then floor is used instead of ceiling}
}
\value{
numeric vector after ceiling()
}
\description{
Rounds up numbers for labels in plots
}
