% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatter plot between prevalence and growth rate}
\usage{
scatter_plot(
  df,
  med,
  x.min = c(0.9, 1.1),
  y.min = c(0.99, 1.01),
  xvar = "confirmed_rate_1M_pop",
  yvar = "growth_factor_3",
  coefflm = NULL,
  addLabels = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{med}{list with median values for x and y}

\item{x.min}{numeric adjustment for Cartesian x axis}

\item{y.min}{numeric adjustment for Cartesian y axis}

\item{xvar}{character variable name for x axis}

\item{yvar}{character variable name for y axis}

\item{coefflm}{numeric, intercept and slop of simple lm model}

\item{addLabels}{logical,if TRUE then labels are added to points}
}
\value{
ggplot plot
}
\description{
scatter plot between prevalence and growth rate
}
