% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatter_plot}
\alias{scatter_plot}
\title{scatterplot between prevalence and growth rate}
\usage{
scatter_plot(df, med, x.min = c(0.9, 1.125), y.min = c(0.99, 1.02))
}
\arguments{
\item{df}{data.frame}

\item{med}{list with median values for x and y}

\item{x.min}{numeric adjustment for cartesian x axis}

\item{y.min}{numeric adjustment for cartesian y axis}
}
\value{
ggplot plot
}
\description{
scatterplot between prevalence and growth rate
}
