% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stackedbarplot_plot}
\alias{stackedbarplot_plot}
\title{stacked barplot status}
\usage{
stackedbarplot_plot(df, color, percent = T, text)
}
\arguments{
\item{df}{data.frame}

\item{color}{string used to define color}

\item{percent}{logical to make the y axis in percent}

\item{y_min}{min value on y axis}
}
\value{
ggplot plot
}
\description{
stacked barplot status
}
