% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stackedbarplot_plot}
\alias{stackedbarplot_plot}
\title{stacked barplot status}
\usage{
stackedbarplot_plot(df, percent = T)
}
\arguments{
\item{df}{data.frame}

\item{percent}{logical to make the y axis in percent}
}
\value{
ggplot plot
}
\description{
stacked barplot status
}
