% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switzerland.R
\name{switzerland}
\alias{switzerland}
\title{Coronavirus COVID-19 data - Switzerland}
\usage{
switzerland(type = "state", raw = FALSE)
}
\arguments{
\item{type}{one of \code{country} (data by country) or \code{state} (data by canton). Default \code{state}, data by canton.}

\item{raw}{logical. Skip data cleaning? Default \code{FALSE}.}
}
\value{
Return of the internal function \code{\link{covid19}}
}
\description{
Tidy format dataset of the 2019 Novel Coronavirus COVID-19 (2019-nCoV) epidemic.
Swiss data by country or state (cantons).
The data are downloaded in real-time, processed and merged with demographic indicators (\code{\link{CH}}).
}
\details{
Data pulled from \href{https://github.com/openZH/covid_19}{Open Government Data} which are communicated
by official Swiss canton's sources.
}
\examples{
# data by country
x <- switzerland("country")

# data by canton
x <- switzerland("state")

}
\seealso{
\code{\link{world}}, \code{\link{diamond}}, \code{\link{italy}}, \code{\link{liechtenstein}}
}
