% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_USA.R
\name{USA}
\alias{USA}
\title{United States}
\source{
https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/iso_USA.R
}
\usage{
USA(level, ...)
}
\description{
Data available at level 1 (nation), level 2 (states), and level 3 (counties).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}
(confirmed cases, recovered, deaths);
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_ourworldindata.org.R}{Our World in Data}
(tests, hospitalizations, vaccines);
\href{https://data.worldbank.org/indicator/SP.POP.TOTL}{World Bank Open Data}
(population 2018).

\bold{Level 2.}
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.nytimes.covid19data.R}{The New York Times}
(confirmed cases, deaths);
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_healthdata.gov.R}{U.S. Department of Health & Human Services}
(tests, hospitalized patients and intensive care);
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_covidtracking.com.R}{The Covid Tracking Project}
(patients requiring ventilation, recovered);
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_ourworldindata.org.R}{Our World in Data}
(vaccines);
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}
(population 2020).

\bold{Level 3.}
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.nytimes.covid19data.R}{The New York Times}
(confirmed cases, deaths);
\href{https://github.com/covid19datahub/COVID19/blob/1ebc0827a18cfb34aacd502a84b53140d2c52054/R/ds_github.cssegisanddata.covid19.R}{Johns Hopkins Center for Systems Science and Engineering}
(population 2020).
}

\concept{level 1}
\concept{level 2}
\concept{level 3}
