% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_USA.R
\name{USA}
\alias{USA}
\title{United States}
\usage{
USA(level, cache)
}
\description{
The data are available at 3 different levels of granularity:
level 1 (nation), level 2 (states), and level 3 (counties).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_jhucsse_git.R}{Johns Hopkins Center for Systems Science and Engineering}
(confirmed cases, recovered, deaths, population);
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_ourworldindata_org.R}{Our World in Data}
(hospitalizations, intensive care, test, vaccines).

\bold{Level 2.}
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_fun_name.R}{Data Source Name}
(variables, separated, by, comma);
\href{https://github.com/covid19datahub/COVID19/blob/master/R/ds_fun_name_2.R}{Data Source Name 2}
(variables, separated, by, comma).
}

\concept{official}
