% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_VIR.R
\name{VIR}
\alias{VIR}
\title{Virgin Islands, U.S.}
\source{
\url{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/iso_VIR.R}
}
\usage{
VIR(level)
}
\description{
Virgin Islands, U.S.
}
\section{Data Sources}{

\subsection{Level 1}{
\itemize{
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/inst/extdata/db/ISO.csv}{World Bank Open Data (2018)}: population.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_github.nytimes.covid19data.R}{The New York Times}:
confirmed cases,
deaths.
\item \href{https://github.com/covid19datahub/COVID19/blob/db2d6d3b0348d924f0771baebf6d102b37eece28/R/ds_ourworldindata.org.R}{Our World in Data}:
tests,
total vaccine doses administered,
people with at least one vaccine dose,
people fully vaccinated,
hospitalizations,
intensive care.
}
}
}

\concept{Level 1}
