% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_VIR.R
\name{VIR}
\alias{VIR}
\title{Virgin Islands, U.S.}
\source{
https://github.com/covid19datahub/COVID19/blob/70dcf69d4e62adebd67e985664a1d170c48b1796/R/iso_VIR.R
}
\usage{
VIR(level, ...)
}
\description{
Data available at level 1 (nation).
}
\section{Data sources}{


\bold{Level 1.}
\href{https://github.com/covid19datahub/COVID19/blob/70dcf69d4e62adebd67e985664a1d170c48b1796/R/ds_nytimes_git.R}{The New York Times}
(confirmed cases, deaths);
\href{https://github.com/covid19datahub/COVID19/blob/70dcf69d4e62adebd67e985664a1d170c48b1796/R/ds_ourworldindata.org.R}{Our World in Data}
(tests, hospitalizations, vaccines);
\href{https://data.worldbank.org/indicator/SP.POP.TOTL}{World Bank Open Data}
(population 2018).
}

\concept{level 1}
