% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WB}
\alias{WB}
\title{Demographic Indicators - Global}
\format{character vector of ids}
\source{
https://data.worldbank.org/
}
\usage{
WB
}
\description{
MEDIAN age: https://www.cia.gov/library/publications/resources/the-world-factbook/fields/343rank.html
}
\keyword{datasets}
